% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{adherance_by_site}
\alias{adherance_by_site}
\title{Adherance by site}
\usage{
adherance_by_site(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include facilitycode, df_date, df_randomized_treatment}
}
\value{
nothing
}
\description{
This function visualizes the treatment crossover or any nonadherance occured during the Tobra 
study.
}
\examples{
\dontrun{
adherance_by_site()
}
}
