% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{adherence_by_id}
\alias{adherence_by_id}
\title{Weight Dearing adherence by ID}
\usage{
adherence_by_id(analytic, random_sample = NULL, facilitycodes = NULL)
}
\arguments{
\item{analytic}{analytic data set that must include study_id, facilitycode, adherence_data (adherence_data must
be a long file with four columns: week, redcap_pt_call_status, text_logs_status, combined_status)}

\item{random_sample}{optional integer to limit to a random sample of IDs}

\item{facilitycodes}{optional character vector to limit to a certain facilities}
}
\value{
An HTML string containing an image tag with the base64-encoded timeline visualization in PNG format.
}
\description{
Creates a timeline visualization for each patient showing adherence recording in text and call logs.
Will most likely work only for Weight Bearing, but potential changes to the adherence_data could make
this work for your study, so please contact an ADS member if you're interested in this visualization.
}
\examples{
adherence_by_id("Replace with Analytic Tibble")
adherence_by_id("Replace with Analytic Tibble", random_sample = 10, facilitycodes = c('AAA', 'AAB'))

}
