% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{adherence_sextant}
\alias{adherence_sextant}
\title{adherence_sextant}
\usage{
adherence_sextant(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include adherence_to_intervention_dwc,
adherence_to_intervention_post_dwc, adherence_to_no_other_antibiotic_dwc}
}
\value{
nothing
}
\description{
This function visualizes the treatment characteristics per protocol and assignmnet for Sextant.
}
\examples{
\dontrun{
adherence_sextant()
}
}
