% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{adjudications_and_discontinuations_by_type}
\alias{adjudications_and_discontinuations_by_type}
\title{Number of Adjudications and Discontinuations by type}
\usage{
adjudications_and_discontinuations_by_type(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include screened; inappropriate_enrollment; 
late_ineligible; late_refusal; withdrawn_patient; withdrawn_physician; adjudication_pending; 
dead; sae_count; protocol_deviation_screen_consent; protocol_deviation_procedural; protocol_deviation_administrative}
}
\value{
An HTML table.
}
\description{
This function visualizes the number of discontinuations, SAEs and Protocol Deviations by type
This was originally made for NSAID
}
\examples{
\dontrun{
adjudications_and_discontinuations_by_type()
}
}
