% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{amputations_and_gustilo_injury_characteristics}
\alias{amputations_and_gustilo_injury_characteristics}
\title{Amputations and gustilo injury characteristics}
\usage{
amputations_and_gustilo_injury_characteristics(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include enrolled, 
injury_gustilo_type, injury_amputation_status}
}
\value{
html table
}
\description{
Visualizes the injury characteristics for the amputation status (by category), 
and the fracture type (by gustilo), for each in the given study.
The returned table has two columns, one for the amputation status 
and gustilo type, and one for their respective counts.
}
\examples{
amputations_and_gustilo_injury_characteristics("Replace with Analytic Tibble")

}
