% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{ankle_and_plateau_x_ray_and_measurement_status}
\alias{ankle_and_plateau_x_ray_and_measurement_status}
\title{Ankle and Plateau X-Ray and Measurement Status}
\usage{
ankle_and_plateau_x_ray_and_measurement_status(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include sixweek_visit_expected, 
threemonth_visit_expected, sixmonth_visit_expected, twelvemonth_visit_expected, injury_type,
radiographs_taken_6wk, radiographs_taken_3mo, radiographs_taken_6mo, radiographs_taken_12mo,
plat_tib_fib_overlap_6mo, plat_sagittal_pl_alignment_6mo, plat_patella_centered_6mo, 
plat_medial_prox_tibia_deg_6mo, plat_medial_lateral_diff_6mo, plat_condylar_width_6mo, plat_art_step_off_medial_6mo, 
plat_art_step_off_lateral_6mo, plat_femur_tibia_deg_6mo,
plat_tib_fib_overlap_3mo, plat_sagittal_pl_alignment_3mo, plat_patella_centered_3mo, 
plat_medial_prox_tibia_deg_3mo, plat_medial_lateral_diff_3mo, plat_condylar_width_3mo, plat_art_step_off_medial_3mo, 
plat_art_step_off_lateral_3mo, plat_femur_tibia_deg_3mo,
plat_tib_fib_overlap_6wk, plat_sagittal_pl_alignment_6wk, plat_patella_centered_6wk, 
plat_medial_prox_tibia_deg_6wk, plat_medial_lateral_diff_6wk, plat_condylar_width_6wk, plat_art_step_off_medial_6wk, 
plat_art_step_off_lateral_6wk, plat_femur_tibia_deg_6wk,
plat_tib_fib_overlap_12mo, plat_sagittal_pl_alignment_12mo, plat_patella_centered_12mo, 
plat_medial_prox_tibia_deg_12mo, plat_medial_lateral_diff_12mo, plat_condylar_width_12mo, plat_art_step_off_medial_12mo, 
plat_art_step_off_lateral_12mo, plat_femur_tibia_deg_12mo
ankle_talar_tilt_degrees_6wk, ankle_sagital_disp_6wk, ankle_coronal_plane_disp_6wk, 
ankle_talar_tilt_degrees_3mo, ankle_sagital_disp_3mo, ankle_coronal_plane_disp_3mo, 
ankle_talar_tilt_degrees_6mo, ankle_sagital_disp_6mo, ankle_coronal_plane_disp_6mo, 
ankle_talar_tilt_degrees_12mo, ankle_sagital_disp_12mo, ankle_coronal_plane_disp_12mo}
}
\value{
nothing
}
\description{
This function visualizes Ankle and Plateau X-Ray and Measurement Status
}
\examples{
\dontrun{
ankle_and_plateau_x_ray_and_measurement_status()
}
}
