% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{ao_gustillo_tscherne_injury_characteristics}
\alias{ao_gustillo_tscherne_injury_characteristics}
\title{AO Gustillo Tscherne Injury Characteristics}
\usage{
ao_gustillo_tscherne_injury_characteristics(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include enrolled, 
injury_gustilo, injury_classification_tscherne, injury_classification_ankle_ao}
}
\value{
nothing
}
\description{
This function visualizes the injury characteristics
}
\examples{
\dontrun{
ao_gustillo_tscherne_characteristics()
}
}
