% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{baseline_characteristics_percent}
\alias{baseline_characteristics_percent}
\title{Baseline Characteristics Percent}
\usage{
baseline_characteristics_percent(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include enrolled, age, age_group, sex, age, race, education, military}
}
\value{
nothing
}
\description{
This function visualizes the categorical percentages of baseline characteristics sex, age, race, education, and military
}
\examples{
\dontrun{
baseline_characteristics_percent()
}
}
