% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{baseline_characteristics_percent_plus}
\alias{baseline_characteristics_percent_plus}
\title{Baseline characteristics percent plus insurance}
\usage{
baseline_characteristics_percent_plus(
  analytic,
  sex = "sex",
  race = "ethnicity_race",
  education = "education_level",
  military = "military_status",
  insurance = "insurance",
  sex_levels = c("Female", "Male", "Missing"),
  race_levels = c("Non-Hispanic White", "Non-Hispanic Black", "Hispanic", "Other",
    "Missing"),
  education_levels = c("Less than High School", "GED or High School Diploma",
    "More than High School", "Refused / Don't know", "Missing"),
  military_levels = c("Active Military", "Active Reserves", "Not Active Duty", "Missing"),
  insurance_levels = c("Yes", "No", "Missing")
)
}
\arguments{
\item{analytic}{analytic data set that must include enrolled, age, age_group, and the constructs specified
in the following parameters.}

\item{sex}{is a meta construct that is required that defaults to "sex"}

\item{race}{is a meta construct that is required that defaults to "ethnicity_race"}

\item{education}{is a meta construct that is required that defaults to "education_level"}

\item{military}{is a meta construct that is required that defaults to "military_status"}

\item{insurance}{is a meta construct that is required that defaults to "insurance"}

\item{sex_levels}{sets default values and orders for sex meta construct}

\item{race_levels}{sets default values and orders for race meta construct}

\item{education_levels}{sets default values and orders for education meta construct}

\item{military_levels}{sets default values and orders for military meta construct}

\item{insurance_levels}{sets default values and orders for insurance meta construct}
}
\value{
html table
}
\description{
Visualizes the categorical distribution of baseline characteristics
sex, age, race, education, military, insurance, enrolled. See below as this is a generic visualization and includes meta construct for each of the
analysis outputs. You may also specify the levels that these outputs have in the function call.
Every categorical characteristic always shows a "Missing" row, including the age_group characteristic,
whose "Missing" row is always ordered last.
Outputs two columns: type (sex, age, race, education, military, insurance), and their respective counts and percentages.
}
\examples{
baseline_characteristics_percent_plus("Replace with Analytic Tibble")
baseline_characteristics_percent_plus("Replace with Analytic Tibble", insurance_levels=c("Yes", "No", "Missing"))
baseline_characteristics_percent_plus("Replace with Analytic Tibble", sex_levels=c("Male", "Female", "Missing"))

}
