% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gray_tables.R
\name{characteristics_treatment}
\alias{characteristics_treatment}
\title{Characteristics of treatment}
\usage{
characteristics_treatment(analytic)
}
\arguments{
\item{analytic}{analytic data set that must study_id, enrolled, df_date, df_plat_surgical_incision, 
df_pil_surgical_incision, df_number_procedures, adherence_to_intervention}
}
\value{
nothing
}
\description{
Visualizes definitive fixation data and study protocol adherence. The function will be formatted 
properly when analytic data includes at most two stages of definitive fixation.

Construct df_number_procedures determines the number of stages of definitive fixation. Constructs
df_plat_surgical_incision and df_pil_surgical_incision use semicolon seperators to count the number 
of incisions at the plateau and pilon sites.
}
\examples{
characteristics_treatment("Replace with Analytic Tibble")

}
