% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_adherence_by_site}
\alias{closed_adherence_by_site}
\title{Closed Adherence by site}
\usage{
closed_adherence_by_site(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include facilitycode, df_date, treatment_arm, treatment_arm}
}
\value{
A kable table
}
\description{
This function visualizes the treatment crossover or any nonadherence occured during the Tobra
study by treatment arm.
}
\examples{
\dontrun{
closed_adherence_by_site(analytic)
}
}
