% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_adherence_sextant}
\alias{closed_adherence_sextant}
\title{Closed adherence_sextant}
\usage{
closed_adherence_sextant(analytic, footnotes = NULL)
}
\arguments{
\item{analytic}{study_id, df_date, enrolled, facilitycode, followup_status_crf08_2wk, followup_status_crf09_2wk}
}
\value{
An HTML table.

Closed followup followup_2wk_status_by_site_tobra

An HTML table.
}
\description{
This function visualizes the treatment characteristics per protocol and assignmnet for Sextant for
each treatment group

This function visualizes 2 weeks followup status by site for Clinical followup form(crf09) and patient
medical record review(crf08) for tobra weekly report
}
\examples{
\dontrun{

closed_adherence_sextant()
}
}
