% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_adjudications_and_discontinuations_by_type}
\alias{closed_adjudications_and_discontinuations_by_type}
\title{Closed Number of Adjudications and Discontinuations by type}
\usage{
closed_adjudications_and_discontinuations_by_type(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include screened, treatment_arm, inappropriate_enrollment;
late_ineligible; late_refusal; withdrawn_patient; withdrawn_physician; adjudication_pending;
dead; sae_count; protocol_deviation_screen_consent; protocol_deviation_procedural; protocol_deviation_administrative}
}
\value{
A kable table
}
\description{
This function visualizes the number of discontinuations, SAEs and Protocol Deviations by type and treatment arm
This was originally made for NSAID
}
\examples{
\dontrun{
closed_adjudications_and_discontinuations_by_type(analytic)
}
}
