% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_amputations_and_gustilo_injury_characteristics}
\alias{closed_amputations_and_gustilo_injury_characteristics}
\title{Amputations and Gustilo injury characteristics, closed version}
\usage{
closed_amputations_and_gustilo_injury_characteristics(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include enrolled, treatment_arm,
injury_gustilo_type, injury_amputation_status}
}
\value{
An HTML table.
}
\description{
This function visualizes the injury characteristics for amputations and Gustilo 
Injury types for Sextant study by treatment arm
}
