% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_ao_gustillo_tscherne_injury_characteristics}
\alias{closed_ao_gustillo_tscherne_injury_characteristics}
\title{Closed AO Gustillo Tscherne Injury Characteristics}
\usage{
closed_ao_gustillo_tscherne_injury_characteristics(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include enrolled, 
injury_gustilo, injury_tscherne, injury_ao_class, treatment_arm}
}
\value{
nothing
}
\description{
This function visualizes the injury characteristics
}
\examples{
\dontrun{
closed_ao_gustillo_tscherne_characteristics()
}
}
