% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_appendix_A_SAEs}
\alias{closed_appendix_A_SAEs}
\title{Appendix A: Listing of Serious Adverse Events(SAEs) for closed report}
\usage{
closed_appendix_A_SAEs(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, sae_data}
}
\value{
HTML.
}
\description{
Prints out formatted data for all the SAEs found in the sae_data long file.
}
\examples{
closed_appendix_A_SAEs("Replace with Analytic Tibble")

}
