% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_appendix_B_deaths}
\alias{closed_appendix_B_deaths}
\title{Appendix B: Listing of any Death for closed report}
\usage{
closed_appendix_B_deaths(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, sae_data, death_date}
}
\value{
An HTML table.
}
\description{
This function visualizes any death occurred during the study time period.
}
