% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{closed_appendix_C_discontinuations}
\alias{closed_appendix_C_discontinuations}
\title{Appendix C: Listing of any Discontinuations for closed report}
\usage{
closed_appendix_C_discontinuations(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, discontinuation_data}
}
\value{
nothing
}
\description{
This function visualizes any discontinuations occurred during the study time period.
}
\examples{
\dontrun{
closed_appendix_C_discontinuations()
}
}
