% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_baseline_characteristics_percent_nm}
\alias{closed_baseline_characteristics_percent_nm}
\title{Closed baseline characteristics percent (no military status)}
\usage{
closed_baseline_characteristics_percent_nm(
  analytic,
  sex = "sex",
  race = "ethnicity_race",
  education = "education_level",
  sex_levels = c("Female", "Male", "Missing"),
  race_levels = c("Non-Hispanic White", "Non-Hispanic Black", "Hispanic", "Other",
    "Missing"),
  education_levels = c("Less than High School", "GED or High School Diploma",
    "More than High School", "Refused / Don't know", "Missing")
)
}
\arguments{
\item{analytic}{analytic dataset that must include enrolled, age, age_group, treatment_arm and all 
the constructs specified in the following construct name parameters}

\item{sex}{name of the construct of the sex characteristic, defaults to "sex"}

\item{race}{name of the construct of the race characteristic, defaults to "ethnicity_race"}

\item{education}{name of the construct of the education characteristic, defaults to "education_level"}

\item{sex_levels}{default values and orders for sex characteristic}

\item{race_levels}{default values and orders for race characteristic}

\item{education_levels}{default values and orders for education characteristic}
}
\value{
An HTML table.
}
\description{
This is the closed version of the baseline_characteristics_percent visualization; see its documentation
for details.
}
\examples{
closed_baseline_characteristics_percent_nm("Replace with Analytic Tibble")
closed_baseline_characteristics_percent_nm("Replace with Analytic Tibble", 
  race_levels=c("Non-Hispanic White", "Non-Hispanic Black", "Hispanic", "Asian / Pacific Islander", "Other", "Missing"))
closed_baseline_characteristics_percent_nm("Replace with Analytic Tibble", 
  sex_levels=c("Male","Female", "Missing"))

}
