% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_baseline_characteristics_percent_plus}
\alias{closed_baseline_characteristics_percent_plus}
\title{Closed Baseline Characteristics Percent Plus Insurance}
\usage{
closed_baseline_characteristics_percent_plus(
  analytic,
  sex = "sex",
  race = "ethnicity_race",
  education = "education_level",
  military = "military_status",
  insurance = "insurance",
  sex_levels = c("Female", "Male", "Missing"),
  race_levels = c("Non-Hispanic White", "Non-Hispanic Black", "Hispanic", "Other",
    "Missing"),
  education_levels = c("Less than High School", "GED or High School Diploma",
    "More than High School", "Refused / Don't know", "Missing"),
  military_levels = c("Active Military", "Active Reserves", "Not Active Duty", "Missing"),
  insurance_levels = c("Yes", "No", "Missing")
)
}
\arguments{
\item{analytic}{This is the analytic data set that must include treatment_arm enrolled, age, age_group}

\item{sex}{is a meta construct that is required that defaults to "sex"}

\item{race}{is a meta construct that is required that defaults to "ethnicity_race"}

\item{education}{is a meta construct that is required that defaults to "education_level"}

\item{military}{is a meta construct that is required that defaults to "military_status"}

\item{insurance}{is a meta construct that is required that defaults to "insurance"}

\item{sex_levels}{sets default values and orders for sex meta construct}

\item{race_levels}{sets default values and orders for race meta construct}

\item{education_levels}{sets default values and orders for education meta construct}

\item{military_levels}{sets default values and orders for military meta construct}

\item{insurance_levels}{sets default values and orders for insurance meta construct}
}
\value{
An HTML table.
}
\description{
This function visualizes the categorical percentages of baseline characteristics sex, age, race, education, military, and insurance.
Every categorical characteristic always shows a "Missing" row, including the age_group characteristic,
whose "Missing" row is always ordered last.

NOTE: This is a closed version of baseline_characteristics_percent_plus, request if baseline_characteristics_percent_plus is present in your study.
See baseline_characteristics_percent_plus for more information.
}
\examples{
closed_baseline_characteristics_percent_plus("Replace with Analytic Tibble")

}
