% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_certification_date_data}
\alias{closed_certification_date_data}
\title{Closed Status of IRB Approvals and Certification by Site}
\usage{
closed_certification_date_data(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include site_certified_date}
}
\value{
An HTML table.
}
\description{
This function returns a list of sites and their dates of
local, DOD, and METRC certifications by treatment arm

NOTE: This is a closed version of certification_date_data, request if certification_date_data is present in your study.
See certification_date_data for more information.
}
\examples{
closed_certification_date_data("Replace with Analytic Tibble")

}
