% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_characteristics_treatment}
\alias{closed_characteristics_treatment}
\title{Closed characteristics of treatment}
\usage{
closed_characteristics_treatment(analytic)
}
\arguments{
\item{analytic}{analytic data set that must study_id, enrolled, df_date, plat_df_surgical_incision, 
pil_df_surgical_incision, df_number_procedures, adherence_to_intervention, treatment_arm}
}
\value{
An HTML table.
}
\description{
Closed version of the characteristics_treatment function; see its documentation for details.
}
\examples{
closed_characteristics_treatment("Replace with Analytic Tibble")

}
