% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_complications_by_severity_relatedness}
\alias{closed_complications_by_severity_relatedness}
\title{Closed Complications by severity and relatedness}
\usage{
closed_complications_by_severity_relatedness(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, complication_data}
}
\value{
An HTML table.
}
\description{
This function visualizes the complications by severity and relatedness for dsmb report
}
\examples{
\dontrun{
closed_complications_by_severity_relatedness()
}
}
