% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_complications_overall}
\alias{closed_complications_overall}
\title{Closed complications reported overall from time_zero}
\usage{
closed_complications_overall(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, complication_data}
}
\value{
A kable table
}
\description{
This function visualizes the number of complications(number of subjects) reported at all time points(Overall) 
from time_zero overall, and by each treatment arm.
}
\examples{
\dontrun{
closed_complications_overall(analytic)
}
}
