% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_complications_within_6mo}
\alias{closed_complications_within_6mo}
\title{Closed complications reported within 6 months from time_zero}
\usage{
closed_complications_within_6mo(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, complication_data, time_zero}
}
\value{
A kable table
}
\description{
This function visualizes the number of complications(number of subjects) reported within 6 months from
time_zero by severity overall, and by each treatment arm.
}
\examples{
\dontrun{
closed_complications_within_6mo(analytic)
}
}
