% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_figures.R
\name{closed_consort_diagram_wb_publication}
\alias{closed_consort_diagram_wb_publication}
\title{Consort Diagram: publication}
\usage{
closed_consort_diagram_wb_publication(analytic)
}
\arguments{
\item{analytic}{analytic data set that must include study_id, screened, ineligible, eligible,
refused, consented, randomized, enrolled, time_zero, adjudicated_discontinued, completed, 
safety_set, exclusive_safety_set, not_completed, not_expected, active, missed_final_followup, incomplete_final_followup}
}
\value{
An HTML string containing an image tag with the base64-encoded consort diagram in PNG format.
}
\description{
The closed version of consort_diagram_wb_publication, breaking down study cancellations by treatment_arm.

This consort diagram was made for the Weight Bearing study, and so is unlikely to work for yours.

For other consort diagrams that may better fit your study, refer to: consort_diagram_no_definitive_event, 
dsmb_consort_diagram, dsmb_consort_diagram_pre_no_def, dsmb_consort_diagram_pre_no_def_shifted_consent, 
dsmb_consort_diagram_pre_shifted_consent, dsmb_nsaid_consort_diagram.
}
