% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_dssi_reported_adjudicated}
\alias{closed_dssi_reported_adjudicated}
\title{Closed Deep Surgical Site Infection reported & adjudicated}
\usage{
closed_dssi_reported_adjudicated(analytic, footnotes = NULL)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, enrolled, followup_expected_6mo, infection_reported_6mo, infection_adjudicated_6mo, 
infection_adjudication_pending_6mo}
}
\value{
A kable table
}
\description{
This function visualizes the treatment crossover or any nonadherence occured during the Tobra
study by treatment arm.
}
\examples{
\dontrun{
closed_dssi_reported_adjudicated(analytic)
}
}
