% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_enrollment_by_site_var_disc_tobra_sextant}
\alias{closed_enrollment_by_site_var_disc_tobra_sextant}
\title{Closed Enrollment by site tobra and sextant (var discontinued)}
\usage{
closed_enrollment_by_site_var_disc_tobra_sextant(analytic, days = 14)
}
\arguments{
\item{analytic}{This is the analytic data set that must include screened, eligible, refused, not_consented, not_randomized, consented_and_randomized, enrolled, site_certified_days,
facilitycode, adjudicated_discontinued, screened_date, treatment_arm}

\item{days}{Number of days to look back for enrollment activity (default is 14)}
}
\value{
A kable table
}
\description{
This function visualizes the number of subjects enrolled, not enrolled etc, with specs for last 14 days and average by week
study by treatment arm.
}
\examples{
\dontrun{
closed_enrollment_by_site_var_disc_tobra_sextant(analytic)
}
}
