% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_enrollment_status_by_site}
\alias{closed_enrollment_status_by_site}
\title{Closed Number of Subjects Screened, Eligible, Enrolled and Not Enrolled}
\usage{
closed_enrollment_status_by_site(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include screened, 
eligible, refused, consented, enrolled, not_consented, discontinued_pre_randomization, site_certified_days, 
facilitycode, late_ineligible, treatment_arm}
}
\value{
A kable table
}
\description{
This function visualizes the enrollment totals for each site by treatment arm
}
\examples{
\dontrun{
closed_enrollment_status_by_site(analytic)
}
}
