% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_enrollment_status_by_site_var_discontinued}
\alias{closed_enrollment_status_by_site_var_discontinued}
\title{Number of Subjects Screened, Eligible, Enrolled and Not Enrolled (Variable Discontinued)}
\usage{
closed_enrollment_status_by_site_var_discontinued(
  analytic,
  discontinued = "discontinued",
  discontinued_colname = "Discontinued",
  only_total = FALSE,
  footnotes = NULL
)
}
\arguments{
\item{analytic}{This is the analytic data set that must include screened, eligible, refused, consented, 
enrolled, not_consented, site_certification_data_metrc_date, facilitycode, treatment_arm}

\item{discontinued}{name of the construct for discontinued status}

\item{discontinued_colname}{column name for discontinued to appear in visualization like "Adjudicated Discontinued"}

\item{only_total}{hide all the site specific rows}
}
\value{
An HTML table.
}
\description{
Closed version of the enrollment_status_by_site_var_discontinued function; see its documentation for
details.
}
\examples{
closed_enrollment_status_by_site_var_discontinued("Replace with Analytic Tibble")
closed_enrollment_status_by_site_var_discontinued("Replace with Analytic Tibble", only_total = TRUE)

}
