% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_enrollment_status_by_site_var_discontinued_i}
\alias{closed_enrollment_status_by_site_var_discontinued_i}
\title{Closed enrollment by site Tobra and Sextant (variable discontinued)}
\usage{
closed_enrollment_status_by_site_var_discontinued_i(
  analytic,
  discontinued = "discontinued",
  discontinued_colname = "Discontinued",
  pre_screened = NULL,
  pre_screened_eligible = NULL,
  only_total = FALSE
)
}
\arguments{
\item{analytic}{analytic data set that must include constructs screened, eligible, refused, consented, 
not_consented, not_randomized, consented_and_randomized, enrolled, site_certification_date, 
facilitycode, screened_date, randomized, consent_date, discontinued, treatment_arm,}

\item{discontinued}{the name of the construct where you can specify your discontinued construct like 
'discontinued' or 'adjudicated_discontinued' (defaults to 'discontinued')}

\item{discontinued_colname}{this determines the label applied to the discontinued column of your 
choosing (defaults to 'Discontinued')}

\item{days}{the number of last days to include in the last days summary section of the table}

\item{include_exclusive_safety_set}{this is a toggle that will include a exclusive_safety_set construct 
if you want it included (defaults to FALSE)}
}
\value{
An HTML table.
}
\description{
This is the closed version of the enrollment_status_by_site_var_discontinued_i function; see its documentation
for details.
}
\examples{
closed_enrollment_status_by_site_var_discontinued_i("Replace with Analytic Tibble")

}
