% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_expected_and_followup_visit_by_site}
\alias{closed_expected_and_followup_visit_by_site}
\title{Expected visit status for 3 Months, 6 Months, and 12 Months followup by EACH SITE}
\usage{
closed_expected_and_followup_visit_by_site(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include facilitycode, followup_due_3mo, 
followup_due_6mo, followup_due_12mo, followup_status_3mo, followup_status_6mo, followup_status_12mo, treatment_arm}
}
\value{
nothing
}
\description{
This function outputs the expected and followup visit status by each site and  uses status 
constructs but treats early, late and complete as mutually exclusive.
Therefore, complete is renamed to "On Time" and all three of them combined to Complete.
}
\examples{
\dontrun{
closed_expected_and_followup_visit_by_site()
}
}
