% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_expected_and_followup_visit_sextant}
\alias{closed_expected_and_followup_visit_sextant}
\title{Closed Expected visit status for 3 Months, 6 Months, and 12 Months followup for Sextant}
\usage{
closed_expected_and_followup_visit_sextant(analytic, footnotes = NULL)
}
\arguments{
\item{analytic}{study_id, df_date, enrolled, facilitycode, followup_status_crf08_3mo, followup_status_crf09_3mo}
}
\value{
nothing
Closed followup followup_3mo_status_by_site_tobra

html table
}
\description{
This function uses status constructs but treats early, late and complete as mutually exclusive.
Therefore, complete is renamed to "On Time" and all three of them combined to Complete.

This function visualizes 6 month followup status by site for Clinical followup form(crf09) and patient
medical record review(crf08) for tobra weekly report
}
\examples{
\dontrun{

closed_expected_and_followup_visit_sextant()
}
}
