% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_expected_and_followup_visit_tobra}
\alias{closed_expected_and_followup_visit_tobra}
\title{Expected visit status for 3 Months, 6 Months, and 12 Months followup for tobra}
\usage{
closed_expected_and_followup_visit_tobra(analytic, footnotes = NULL)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, time_zero, followup_status_crf09_2wk, followup_status_crf09_3mo, followup_status_crf09_6mo, 
followup_status_crf09_12mo}
}
\value{
nothing
}
\description{
This function onlt looks at the clinical followup form (CRF09) for tobra, designations may have to be looked at
}
\examples{
\dontrun{
closed_expected_and_followup_visit_tobra()
}
}
