% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_expected_visits_by_followup_period}
\alias{closed_expected_visits_by_followup_period}
\title{Closed Number of Visits Expected, Completed, Missed, and out of Window}
\usage{
closed_expected_visits_by_followup_period(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include enrolled, treatment_arm, followup_complete_4wk_6wk, followup_due_4wk_6wk, followup_early_4wk_6wk, followup_incomplete_4wk_6wk,
followup_late_4wk_6wk, followup_missing_4wk_6wk, followup_not_started_4wk_6wk, followup_ontime_4wk_6wk,
followup_complete_7wk_9wk, followup_due_7wk_9wk, followup_early_7wk_9wk, followup_incomplete_7wk_9wk,
followup_late_7wk_9wk, followup_missing_7wk_9wk, followup_not_started_7wk_9wk, followup_complete_12wk_16wk,
followup_due_12wk_16wk, followup_early_12wk_16wk, followup_incomplete_12wk_16wk, followup_late_12wk_16wk,
followup_missing_12wk_16wk, followup_not_started_12wk_16wk, followup_ontime_12wk_16wk, followup_complete_22wk_32wk,
followup_due_22wk_32wk, followup_early_22wk_32wk, followup_incomplete_22wk_32wk, followup_late_22wk_32wk,
followup_missing_22wk_32wk, followup_not_started_22wk_32wk, followup_ontime_22wk_32wk}
}
\value{
A kable table
}
\description{
This function visualizes the expected visits, and which of those are completed in what relative window,
and which of those are missed by treatment arm
}
\examples{
\dontrun{
closed_expected_visits_by_followup_period(analytic)
}
}
