% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_followup_12mo_status_by_site_sextant}
\alias{closed_followup_12mo_status_by_site_sextant}
\title{Closed Followup 12 month status by site for Sextant}
\usage{
closed_followup_12mo_status_by_site_sextant(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, eligible, enrolled, facilitycode,
followup_expected_12mo, time_zero, followup_complete_crf12_12mo, followup_incomplete_crf12_12mo,
followup_early_crf12_12mo, followup_late_crf12_12mo, followup_missing_crf12_12mo,
followup_not_started_crf12_12mo, followup_status_crf14_crf15_12mo, followup_status_crf09_12mo, treatment_arm}
}
\value{
A kable table
}
\description{
This function visualizes 12 month followup status by site using CRF12(Clinical followup form),
PROMIS pain interference/intensity + CRF15(Survey version) or CRF14 + CRF15(Redcap version), and CRF08(Medical
Record Review) form by treatment arm
}
\examples{
\dontrun{
closed_followup_12mo_status_by_site_sextant(analytic)
}
}
