% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_followup_3mo_status_by_site_tobra}
\alias{closed_followup_3mo_status_by_site_tobra}
\title{Closed followup followup_3mo_status_by_site_tobra}
\usage{
closed_followup_3mo_status_by_site_tobra(analytic, footnotes = NULL)
}
\arguments{
\item{analytic}{study_id, df_date, enrolled, facilitycode, followup_status_crf08_3mo, followup_status_crf09_3mo}
}
\value{
html table
}
\description{
This function visualizes 6 month followup status by site for Clinical followup form(crf09) and patient
medical record review(crf08) for tobra weekly report
}
\examples{
\dontrun{
closed_followup_3mo_status_by_site_tobra()
}
}
