% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_followup_6mo_status_by_site_sextant}
\alias{closed_followup_6mo_status_by_site_sextant}
\title{Closed Followup 6 month status by site for Sextant}
\usage{
closed_followup_6mo_status_by_site_sextant(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, eligible, enrolled, facilitycode,
followup_expected_6mo, time_zero, followup_complete_crf12_6mo, followup_incomplete_crf12_6mo,
followup_early_crf12_6mo, followup_late_crf12_6mo, followup_missing_crf12_6mo, followup_not_started_crf12_6mo,
followup_status_crf14_crf15_6mo, treatment_arm}
}
\value{
A kable table
}
\description{
This function visualizes 6 month followup status by site for Clinical followup form(crf12) and patient
reported outcome forms(CRF 14 & 15) for Sextant weekly report by treatment arm
}
\examples{
\dontrun{
closed_followup_6mo_status_by_site_sextant(analytic)
}
}
