% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_followup_form_all_timepoints_by_site}
\alias{closed_followup_form_all_timepoints_by_site}
\title{Closed Followup Data Single Form All Timepoints By Site}
\usage{
closed_followup_form_all_timepoints_by_site(
  analytic,
  form_selection = "Overall"
)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, followup_data}
}
\value{
nothing
}
\description{
Returns the designated followup form status by site, for all timepoints,
separated by treatment_arm
}
\examples{
\dontrun{
closed-followup_form_all_timepoints_by_site()
}
}
