% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_followup_form_at_timepoint_by_site}
\alias{closed_followup_form_at_timepoint_by_site}
\title{Closed Followup Data Single Form and Timepoint By Site}
\usage{
closed_followup_form_at_timepoint_by_site(
  analytic,
  timepoint,
  form_selection,
  name = NULL
)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, followup_data}
}
\value{
nothing
}
\description{
Returns the designated followup form status by site, by treatment data
}
\examples{
\dontrun{
closed_followup_form_at_timepoint_by_site()
}
}
