% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_followup_forms_all_timepoints}
\alias{closed_followup_forms_all_timepoints}
\title{Closed Followup Data Multiple Forms and All Timepoints}
\usage{
closed_followup_forms_all_timepoints(analytic, forms = NULL, timepoints = NULL)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, followup_data, treatment_arm}

\item{forms}{followup forms to output, as found in the followup_data construct}

\item{timepoints}{timepoints to output, as found in the followup_data construct}
}
\value{
An HTML table.
}
\description{
Returns the designated followup forms status at a specified range
of timepoints, not specified by site, separated vertically by treatment_arm
}
\examples{
\dontrun{
closed_followup_forms_all_timepoints()
}
}
