% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_followup_forms_at_timepoint_by_site}
\alias{closed_followup_forms_at_timepoint_by_site}
\title{Closed Followup Data Multiple Forms and Single Timepoint By Site}
\usage{
closed_followup_forms_at_timepoint_by_site(
  analytic,
  timepoint,
  forms,
  names = NULL
)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, followup_data}

\item{timepoint}{the point in time to be considered in the visualization}

\item{forms}{the form to be considered in the visualization}

\item{names}{This is the analytic data set that must include study_id, followup_data}
}
\value{
html table
}
\description{
Visualizes the designated followup form status by site.

This is a closed version of followup_forms_at_timepoint_by_site, request if followup_forms_at_timepoint_by_site is present in your study. 
See followup_forms_at_timepoint_by_site for more information.
}
\examples{
closed_followup_forms_at_timepoint_by_site("Replace with Analytic Tibble", "3 Month", c("Form 2", "Form 4"))

}
