% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_fracture_characteristics}
\alias{closed_fracture_characteristics}
\title{Closed Fracture Characteristics}
\usage{
closed_fracture_characteristics(analytic)
}
\arguments{
\item{analytic}{analytic data set that must include study_id, enrolled, fracture_type, injury_gustilo,
treatment_arm}
}
\value{
An HTML table.
}
\description{
This is the closed version of the fracture_characteristics function; see its documentation for details.
}
\examples{
closed_fracture_characteristics("Replace with Analytic Tibble")

}
