% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_ineligibility_by_reasons}
\alias{closed_ineligibility_by_reasons}
\title{Closed Number of patients Ineligible by Top 5 reasons of Exclusion}
\usage{
closed_ineligibility_by_reasons(analytic, n_top_reasons = 5)
}
\arguments{
\item{analytic}{This is the analytic data set that must include facilitycode,  screened, ineligible, ineligibility_reasons, treatment_arm}
}
\value{
A kable table
}
\description{
This function visualizes the number of patients Ineligible by Top 5 reasons of Exclusion criteria by treatment arm
}
\examples{
\dontrun{
closed_ineligibility_by_reasons(analytic)
}
}
