% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_injury_ankle_plateau_characteristics}
\alias{closed_injury_ankle_plateau_characteristics}
\title{Closed Injury characteristics for OTA classification and Schatzker Type injuries}
\usage{
closed_injury_ankle_plateau_characteristics(analytic)
}
\arguments{
\item{analytic}{analytic data set that must include treatment_arm, injury_type, injury_classification_ankle_ota, 
injury_classification_plat_schatzker, enrolled}
}
\value{
An HTML table.
}
\description{
This is the closed version of the injury_ankle_plateau_characteristics function; see its documentation
for details
}
\examples{
\dontrun{
closed_injury_ankle_plateau_characteristics()
}
}
