% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_nonunion_surgery_outcome}
\alias{closed_nonunion_surgery_outcome}
\title{Closed Nonunion surgery outcome}
\usage{
closed_nonunion_surgery_outcome(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include enrolled, 
followup_expected_3mo, followup_expected_12mo, nonunion_90days,  nonunion_1yr, treatment_arm}
}
\value{
nothing
}
\description{
This function visualizes the Nonunion surgery outcome
}
\examples{
\dontrun{
closed_nonunion_surgery_outcome()
}
}
