% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_not_complete_sae_deviation_by_type}
\alias{closed_not_complete_sae_deviation_by_type}
\title{Closed Number of Discontinued Participants, SAEs, and Protocol Deviations by type}
\usage{
closed_not_complete_sae_deviation_by_type(
  analytic,
  include_ae = FALSE,
  factors = list()
)
}
\arguments{
\item{analytic}{analytic data set that must include enrolled, not_expected_reason, not_completed_reason,
protocol_deviation_screen_consent, protocol_deviation_procedural, protocol_deviation_administrative, 
sae_count, treatment_arm}

\item{include_ae}{boolean to include adverse events in the table, defaults to FALSE}

\item{factors}{dynamic argument to set a custom order in the final visualization, accepted names are 
not_completed, not_espected, deviation_a, deviation_p, deviation_sc and the values are character vectors
of final strings in desired order.}
}
\value{
An HTML table.
}
\description{
This is the closed version of the not_complete_sae_deviation_by_type function; see its documentation
for details.
}
\examples{
closed_not_complete_sae_deviation_by_type("Replace with Analytic Tibble")

}
