% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_not_complete_sae_deviation_by_type_auto_categories}
\alias{closed_not_complete_sae_deviation_by_type_auto_categories}
\title{Closed Number of Non-Completing Participants, SAEs, and Protocol Deviations by type}
\usage{
closed_not_complete_sae_deviation_by_type_auto_categories(
  analytic,
  category_defaults = c("Safety", "Informed Consent", "Eligibility",
    "Protocol Implementation", "Other")
)
}
\arguments{
\item{analytic}{This is the analytic data set that must include enrolled, not_expected_reason, 
not_completed_reason, protocol_deviation_full_data, sae_count}
}
\value{
An HTML table.
}
\description{
This function visualizes the number of discontinuations, SAEs and Protocol Deviations by type
Now with AUTO Protocol Deviation Categorization!
}
\examples{
\dontrun{
closed_not_complete_sae_deviation_by_type_auto_categories()
}
}
