% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_not_enrolled_for_other_reasons}
\alias{closed_not_enrolled_for_other_reasons}
\title{Not enrolled for other reasons}
\usage{
closed_not_enrolled_for_other_reasons(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, facilitycode, able_to_participate,
nonparticipation_text_given, constraint_noconsent, constraint_admin, constraint_othr, constraint_othr_txt, screened, treatment_arm}
}
\value{
A kable table
}
\description{
This function visualizes list of study_ids who were screened howevere were not enrolled for "Other"
reasons by treatment arm
}
\examples{
\dontrun{
closed_not_enrolled_for_other_reasons(analytic)
}
}
