% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_other_reason_refusal_by_site}
\alias{closed_other_reason_refusal_by_site}
\title{Closed Other reason of refusal by each site}
\usage{
closed_other_reason_refusal_by_site(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, facilitycode, screened_date,
refused_reason_other, treatment_arm}
}
\value{
A kable table
}
\description{
This function visualizes list of each "Other" reason of refusal, total screened by each site and treatment arm.
}
\examples{
\dontrun{
closed_other_reason_refusal_by_site(analytic)
}
}
