% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_overall_complications}
\alias{closed_overall_complications}
\title{Overall complications closed}
\usage{
closed_overall_complications(
  analytic,
  relatedness = TRUE,
  WB = NULL,
  breakout_other = FALSE
)
}
\arguments{
\item{analytic}{analytic data set that must include study_id, treatment_arm, complication_data}

\item{relatedness}{includes that column}

\item{WB}{if the study is Weight Bearing}

\item{breakout_other}{boolean. If TRUE, replaces "Other" with "Other: [other_info]". Defaults to FALSE.}
}
\value{
html table
}
\description{
Returns a table of the overall complications separated by Treatment Arm (Group A vs Group B).
Ordered by complication alphabetically (Other at bottom), then by relatedness, then by severity.
}
