% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_pathogen_characteristics}
\alias{closed_pathogen_characteristics}
\title{Closed pathogen Characteristics}
\usage{
closed_pathogen_characteristics(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include dssi_data}
}
\value{
An HTML table styled with kableExtra.
}
\description{
Closed version of pathogen_characteristics, see that function for details.
}
\examples{
\dontrun{
}
}
