% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_promis_stats_by_time}
\alias{closed_promis_stats_by_time}
\title{closed PROMIS stats by time}
\usage{
closed_promis_stats_by_time(analytic)
}
\arguments{
\item{analytic}{enrolled, promis_pf 6wk - 12mo constructs, promis_pain_interference 6wk - 12mo constructs}
}
\value{
An HTML table.
}
\description{
Returns stat data on promis scores, including sample size, mean, and standard deviation. This is the
closed_version of promis_stats_by_time
}
\examples{
wbs_main_paper_aos("Replace with Analytic Tibble")

}
