% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_refusal_reasons_by_site}
\alias{closed_refusal_reasons_by_site}
\title{Closed Refusal reasons by each site}
\usage{
closed_refusal_reasons_by_site(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include facilitycode, screened, refused, refused_reason, treatment_arm}
}
\value{
A kable table
}
\description{
This function visualizes the reasons of refusal, total refused and total screened by each site and treatment arm.
}
\examples{
\dontrun{
closed_refusal_reasons_by_site(analytic)
}
}
