% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_survival_analysis_kaplan_meier}
\alias{closed_survival_analysis_kaplan_meier}
\title{Closed Survival Analysis Kaplan-Meier}
\usage{
closed_survival_analysis_kaplan_meier(
  analytic,
  type_construct,
  days_construct,
  outcome_length,
  pre_filter_constructs = NULL,
  remove_zero_day_events = TRUE,
  arm_labels = c(`0` = "Control", `1` = "Treatment"),
  outcome_label = "Outcome"
)
}
\arguments{
\item{analytic}{This is the analytic dataset that must include enrolled, treatment_arm}

\item{type_construct}{the name of the column of the analytic dataset that must include whether the outcome for that participant was a check or event}

\item{days_construct}{the name of the column of the analytic dataset that must include the number of days till check or event}

\item{outcome_length}{number of days for this outcome}

\item{remove_zero_day_events}{defaults to TRUE removes zero day events to match STATA behavior}

\item{arm_labels}{named chr vec, c("0" = "Early Weight bearing","1" = "Restricted Weight bearing")}

\item{pre_filter_construct}{defaults to NULL but can be used to filter the participants using the name or names of other columns of the analytic dataset}
}
\value{
An HTML table.
}
\description{
This function outputs a table with the specified injury characteristics and patient characteristics
for enrolled patients with "Ankle" injuries. This table is produced for Weight bearing main paper.
}
\examples{
closed_survival_analysis_kaplan_meier("Replace with Analytic Tibble")

}
