% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_treatment_characteristics}
\alias{closed_treatment_characteristics}
\title{Closed treatment_characteristics (var discontinued)}
\usage{
closed_treatment_characteristics(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, 
enrolled, df_date, plat_df_surgical_incision, pil_df_surgical_incision,
df_number_procedures, df_randomized_treatment,
injury_classification_plat_ao, df_surg_staged, df_surg_n_forms,
randomized, fracture_type, treatment_arm}
}
\value{

}
\description{
This function visualizes the characteristics of the definitive 
fixation process. Data included is completeness, number of df stages, 
statistics on distribution, incision location counts and distribution,
and adherence by treatment arm
}
