% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_treatment_characteristics_sextant}
\alias{closed_treatment_characteristics_sextant}
\title{Closed Treatment characteristics Sextant}
\usage{
closed_treatment_characteristics_sextant(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include local_antibiotic_at_dwc, treatment_arm,
systemic_antibiotic_post_dwc, no_other_antibiotic_at_dwc}
}
\value{
A kable table
}
\description{
This function visualizes the treatment characteristics per protocol and assignment for Sextant by treatment arm.
}
\examples{
\dontrun{
closed_treatment_characteristics_sextant(analytic)
}
}
