% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_treatment_crossover_and_nonadherence}
\alias{closed_treatment_crossover_and_nonadherence}
\title{Closed Treatment crossover and nonadherence}
\usage{
closed_treatment_crossover_and_nonadherence(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include facilitycode, eligible, enrolled, time_zero,
local_antibiotic_at_dwc}
}
\value{
A kable table
}
\description{
This function visualizes the treatment crossover or any nonadherence occured during the Sextant
study by treatment arm.
}
\examples{
\dontrun{
closed_treatment_crossover_and_nonadherence(analytic)
}
}
