% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_visit_status_for_followup_by_form}
\alias{closed_visit_status_for_followup_by_form}
\title{Closed Expected, completed, missing, out of window visits by each form}
\usage{
closed_visit_status_for_followup_by_form(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include treatment_arm mrr_status_6wk, mrr_status_3mo, mrr_status_6mo, 
mrr_status_12mo, cfu_status_6wk, cfu_status_3mo, cfu_status_6mo, cfu_status_12mo, pfu_status_6wk, pfu_status_3mo, 
pfu_status_6mo, pfu_status_12mo, bpi_status_6wk, bpi_status_3mo, bpi_status_6mo,bpi_status_12mo, aos_status_6wk, 
aos_status_3mo, aos_status_6mo, aos_status_12mo, koos_status_6wk, koos_status_3mo, koos_status_6mo, koos_status_12mo, 
followup_expected_6wk, followup_expected_3mo, followup_expected_6mo, followup_expected_12mo, injury_type}
}
\value{
nothing
}
\description{
This function visualizes the expected visits for each timepoint for MRR, CFU, PFU, BPI, AOS, KOOS forms
}
\examples{
\dontrun{
closed_visit_status_for_followup_by_form()
}
}
