% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_wbs_main_paper_all_characteristics}
\alias{closed_wbs_main_paper_all_characteristics}
\title{Closed Weight Bearing ALL characteristics for Main paper}
\usage{
closed_wbs_main_paper_all_characteristics(analytic)
}
\arguments{
\item{analytic}{This is the analytic dataset that must include enrolled, injury_type,
sex, age, ethnicity_race, education_level,
preinjury_productive_activity, preinjury_work_demand, preinjury_work_hours,
tobacco_use, bmi, preinjury_health, insurance_type,
injury_gustilo, injury_classification_ankle_ota, soft_tissue_closure,
injury_mechanism, injury_randomization_days, pre_randomization_mobilization, treatment_arm}
}
\value{
An HTML table.
}
\description{
Closed version of wbs_main_paper_all_characteristics; see it's documentation for details.
}
\examples{
closed_wbs_main_paper_all_characteristics("Replace with Analytic Tibble")

}
