% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_wbs_main_paper_aos}
\alias{closed_wbs_main_paper_aos}
\title{Closed Weight Bearing AOS for Main Paper}
\usage{
closed_wbs_main_paper_aos(analytic)
}
\arguments{
\item{analytic}{enrolled, 
aos_disability_score_injured_leg_12mo, aos_disability_score_injured_leg_3mo, aos_disability_score_injured_leg_6mo, aos_disability_score_injured_leg_6wk, 
aos_pain_score_injured_leg_12mo, aos_pain_score_injured_leg_3mo, aos_pain_score_injured_leg_6mo, aos_pain_score_injured_leg_6wk, 
aos_score_injured_leg_12mo, aos_score_injured_leg_3mo, aos_score_injured_leg_6mo, aos_score_injured_leg_6wk,
treatment_arm}
}
\value{
An HTML table.
}
\description{
Closed version of wbs_main_paper_aos; see it's documentation for details.
}
\examples{
closed_wbs_main_paper_aos("Replace with Analytic Tibble")

}
