% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_wbs_main_paper_bpi}
\alias{closed_wbs_main_paper_bpi}
\title{Closed Weight Bearing BPI for Main Paper}
\usage{
closed_wbs_main_paper_bpi(analytic)
}
\arguments{
\item{analytic}{enrolled, 
bpi_severity_score_6wk, bpi_severity_score_3mo, bpi_severity_score_6mo, bpi_severity_score_12mo, 
bpi_interference_score_6wk, bpi_interference_score_3mo, bpi_interference_score_6mo, bpi_interference_score_12mo, treatment_arm}
}
\value{
An HTML table.
}
\description{
Closed version of wbs_main_paper_bpi; see it's documentation for details.
}
\examples{
closed_wbs_main_paper_bpi("Replace with Analytic Tibble")

}
