% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{consort_diagram}
\alias{consort_diagram}
\title{Consort Diagram}
\usage{
consort_diagram(
  analytic,
  final_period = "12 Month",
  definitive_event = "Definitive Fixation Complete",
  not_expected_adjudicated = TRUE,
  not_enrolled_missing_df_reason = FALSE
)
}
\arguments{
\item{analytic}{analytic data set that must include study_id, screened, ineligible, eligible,
refused, consented, randomized, enrolled, time_zero, adjudicated_discontinued, completed, 
safety_set, exclusive_safety_set, not_completed, not_expected, active, missed_final_followup, incomplete_final_followup}

\item{final_period}{visual label of period of study completion, defaults to "12 Month"}

\item{definitive_event}{visual label of definitive event, defaults to "Definitive Fixation Complete" 
(attached to the count of the df_complete field)}

\item{not_expected_adjudicated}{whether to note that the Not Expected was adjudicated, purely visual
change}

\item{not_enrolled_missing_df_reason}{adds reasons for not enrollment beyond adjudication to consort}
}
\value{
An HTML string containing an image tag with the base64-encoded consort diagram in PNG format.
}
\description{
Visualizes the categorical percentages of study status as well as followup completions. 
Consort diagrams are almost fully customizable in their implementation. 

For other consort diagrams that may better fit your study, refer to: consort_diagram_no_definitive_event, 
dsmb_consort_diagram, dsmb_consort_diagram_pre_no_def, dsmb_consort_diagram_pre_no_def_shifted_consent, 
dsmb_consort_diagram_pre_shifted_consent, dsmb_nsaid_consort_diagram.
}
\examples{
consort_diagram("Replace with Analytic Tibble")

}
