% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{consort_diagram_no_definitive_event}
\alias{consort_diagram_no_definitive_event}
\title{Consort Diagram No Definitive Event}
\usage{
consort_diagram_no_definitive_event(
  analytic,
  final_period = "12 Month",
  not_expected_adjudicated = TRUE
)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, screened, ineligible, eligible,
refused, consented, randomized, enrolled, adjudicated_discontinued, completed, 
safety_set, exclusive_safety_set, not_completed, not_expected, active, missed_final_followup, incomplete_final_followup}

\item{final_period}{Defaults to 12 Month}

\item{not_expected_adjudicated}{whether to note that the Not Expected was adjudicated}
}
\value{
An HTML string containing an image tag with the base64-encoded consort diagram in PNG format.
}
\description{
This function visualizes the categorical percentages of study status as well as followup completions for studies that do not include a definitive event
}
\examples{
\dontrun{
consort_diagram_no_definitive_event()
}
}
