% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{consort_diagram_wb_publication}
\alias{consort_diagram_wb_publication}
\title{Consort Diagram: Weight Bearing publication}
\usage{
consort_diagram_wb_publication(analytic)
}
\arguments{
\item{analytic}{analytic data set that must include 
study_id, screened, ineligible, ineligibility_reasons, refused, constraint_other, constraint_other_txt, consented, 
discontinued_pre_randomization, injury_type (must have ankle and plateau values), randomized, 
late_ineligible, per_protocol_sample, enrolled, consent_date, death_date, withdraw_date,
preinjury_work_status, followup_expected_12mo, completed, outcome_data}
}
\value{
An HTML string containing an image tag with the base64-encoded consort diagram in PNG format.
}
\description{
Visualizes the categorical percentages of study status as well as followup completions. 
Consort diagrams are almost fully customizable in their implementation. 

This consort diagram was made for the Weight Bearing study, and so is unlikely to work for yours.

For other consort diagrams that may better fit your study, refer to: consort_diagram_no_definitive_event, 
dsmb_consort_diagram, dsmb_consort_diagram_pre_no_def, dsmb_consort_diagram_pre_no_def_shifted_consent, 
dsmb_consort_diagram_pre_shifted_consent, dsmb_nsaid_consort_diagram.
}
\examples{
consort_diagram_wb_publication("Replace with Analytic Tibble")

}
