% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{constented_enrolled_reg_tally}
\alias{constented_enrolled_reg_tally}
\title{Race-Ethnicity-Gender Tally's for Enrolled and Consented Participants}
\usage{
constented_enrolled_reg_tally(analytic, last_review_dt, initial_approval_dt)
}
\arguments{
\item{analytic}{analytic data set that must include, enrolled, consented, consent_date, race, ethnicity, and sex}

\item{last_review_dt}{pass in the YYYY:MM:dd date time for the last review of this study {MUST HAVE}}

\item{initial_approval_dt}{pass in the YYYY:MM:dd date time for the last review of this study {MUST HAVE}}
}
\value{
An HTML table.
}
\description{
Tallys up the Races, Etnicities, and Genders for participants who are consented/enrolled since the last review
and participants who are consented/enrolled since the initial approval
}
\examples{
constented_enrolled_reg_tally(analytic, last_review_dt = "2024-06-01", initial_approval_dt = "2024-05-20")

}
