% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{cumulative_enrolled_los}
\alias{cumulative_enrolled_los}
\title{Cumulative enrollment for Length of Stay for NSAID}
\usage{
cumulative_enrolled_los(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, ih_los_days}
}
\value{
An HTML string containing an image tag with the base64-encoded consort diagram in PNG format.
}
\description{
This function visualizes the Cumulative number of patients enrolled for each bucket of number
of days stay during the admission
}
\examples{
cumulative_enrolled_los("Replace with Analytic Tibble")

}
