% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{cumulative_enrollment_goals}
\alias{cumulative_enrollment_goals}
\title{Cumulative Enrollment with Goals}
\usage{
cumulative_enrollment_goals(analytic, start_date, end_date, participant_goal)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, enrolled, consent_date.}

\item{start_date}{The start date for the analysis.}

\item{end_date}{The end date for the analysis.}

\item{participant_goal}{The goal number of participants for the study.}
}
\value{
An HTML string containing an image tag with the base64-encoded consort diagram in PNG format. (intended for plotting)
}
\description{
This function visualizes the cumulative number of patients enrolled, accompanied by 
a participant goal curve which demonstrates the goal relative to the actual cumulative participant enrollment process, 
and a second curve which demonstrates the necessary growth in order to meet that goal, all within a specified start and end date..
}
\examples{
cumulative_enrollment_goals("Replace with Analytic Tibble")

}
