% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{discontinuation_sae_deviation_by_type}
\alias{discontinuation_sae_deviation_by_type}
\title{Number of Discontinued Participants, SAEs, and Protocol Deviations by type}
\usage{
discontinuation_sae_deviation_by_type(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include enrolled, censored_reason, 
protocol_deviation_screen_consent, protocol_deviation_procedural, protocol_deviation_administrative, sae_count}
}
\value{
nothing
}
\description{
This function visualizes the number of discontinuations, SAEs and Protocol Deviations by type
This was originally made for Union
}
\examples{
\dontrun{
discontinuation_sae_deviation_by_type()
}
}
