% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{dsmb_consort_diagram}
\alias{dsmb_consort_diagram}
\title{DSMB Consort Diagram}
\usage{
dsmb_consort_diagram(
  analytic,
  active,
  completed,
  not_enrolled_other,
  completed_str = "Completed 12-month visit"
)
}
\arguments{
\item{analytic}{This is the analytic data set that must include screened, eligible, 
consented, randomized, discontinued, ineligible, refused, late_ineligible, and the meta construct columns}

\item{active}{is a meta construct that is required}

\item{completed}{is a meta construct that is required}

\item{not_enrolled_other}{is a meta construct that is required}

\item{completed_str}{is the text for the completed box that defaults to 'Completed 12-month visit'}
}
\value{
nothing
}
\description{
This function visualizes the categorical percentages of baseline characteristics sex, age, race, education, and military
}
\examples{
\dontrun{
dsmb_consort_diagram()
}
}
