% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{dsmb_consort_diagram_pre_no_def_shifted_consent}
\alias{dsmb_consort_diagram_pre_no_def_shifted_consent}
\title{DSMB Consort Diagram With Pre Screened and No Definitive Event}
\usage{
dsmb_consort_diagram_pre_no_def_shifted_consent(
  analytic,
  final_period = "12 Month",
  adjudicated = FALSE
)
}
\arguments{
\item{analytic}{analytic data set that must include pre_screened, pre_eligible, screened, eligible,
consented, not_consented, randomized, enrolled, refused, completed, not_completed, not_expected, active, missed_final_followup, incomplete_final_followup}

\item{final_period}{labels the final follow-up period box, defaults to "12 Month"}

\item{adjudicated}{whether to use construct adjudicated_discontinued instead of discontinued and 
labels it as such}
}
\value{
An HTML string containing an image tag with the base64-encoded consort diagram in PNG format.
}
\description{
This function visualizes the categorical percentages of Study Status

For other consort diagrams that may better fit your study, refer to: consort_diagram, consort_diagram_no_definitive_event, 
dsmb_consort_diagram, dsmb_consort_diagram_pre_no_def, dsmb_consort_diagram_pre_no_def_shifted_consent, 
dsmb_consort_diagram_pre_shifted_consent, dsmb_nsaid_consort_diagram.
}
\examples{
dsmb_consort_diagram_pre_no_def_shifted_consent("Replace with Analytic Tibble")

}
