% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{dsmb_consort_diagram_pre_post_consent_check}
\alias{dsmb_consort_diagram_pre_post_consent_check}
\title{DSMB consort diagram with pre-screening and no definitive event}
\usage{
dsmb_consort_diagram_pre_post_consent_check(
  analytic,
  missing_post_consent_check,
  missing_post_consent_check_label,
  post_consent_check,
  post_consent_check_label,
  post_consent_eligible,
  post_consent_eligible_label,
  post_consent_ineligible,
  post_consent_ineligible_label,
  final_period = "12 Month",
  adjudicated = FALSE
)
}
\arguments{
\item{analytic}{This is the analytic data set that must include pre_screened, pre_eligible, pre_ineligible, 
screened, eligible, ineligible, consented, not_consented, randomized, enrolled, refused, completed, 
not_completed, not_expected, active, missed_final_followup, incomplete_final_followup}

\item{missing_post_consent_check}{construct name for missing post event check}

\item{missing_post_consent_check_label}{label for post missing event check}

\item{post_consent_check}{construct name for post event check}

\item{post_consent_check_label}{label for post event check}

\item{post_consent_eligible}{construct name for post event eligible}

\item{post_consent_eligible_label}{label for post event eligible}

\item{post_consent_ineligible}{construct name for post event ineligible}

\item{post_consent_ineligible_label}{label for post event ineligible}

\item{final_period}{label of the final period}

\item{adjudicated}{visual option to say that discontinuation was adjudicated, defaults to false}
}
\value{
An HTML string containing an image tag with the base64-encoded consort diagram in PNG format.
}
\description{
Visualizes the counts of different study statuses, including prescreening statuses. The diagram works
from the final_followup constructs rather than any df ones.

See dsmb_consort_diagram and consort_diagram for similar options.
}
\examples{
dsmb_consort_diagram_pre_post_consent_check("Replace with Analytic Tibble", "missing_double_check", "Missing Double Check", "post_consent_check", "Post Consent Check", "post_consent_eligible", "Post Consent Eligible", "post_consent_ineligible", "Post Consent Ineligible", final_period = '3 Month', adjudicated = TRUE)

}
