% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{dsmb_consort_diagram_pre_shifted_consent}
\alias{dsmb_consort_diagram_pre_shifted_consent}
\title{DSMB Consort Diagram With Pre Screened and No Definitive Event and with the Consented Group moved up}
\usage{
dsmb_consort_diagram_pre_shifted_consent(
  analytic,
  final_period = "12 Month",
  adjudicated = FALSE,
  definitive_event = "Nerve Surgery"
)
}
\arguments{
\item{analytic}{This is the analytic data set that must include pre_screened, pre_eligible, screened, eligible,
consented, not_consented, randomized, enrolled, refused, completed, not_completed, not_expected, active, missed_final_followup, incomplete_final_followup}

\item{final_period}{Defaults to 12 Month}

\item{adjudicated}{whether to use adjudicated discontinued and not expected}

\item{definitive_event}{the definitive event}
}
\value{
An HTML string containing an image tag with the base64-encoded consort diagram in PNG format.
}
\description{
This function visualizes the categorical percentages of Study Status
}
\examples{
dsmb_consort_diagram_pre_shifted_consent("Replace with Analytic Tibble")

}
