% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{dsmb_nsaid_consort_diagram}
\alias{dsmb_nsaid_consort_diagram}
\title{DSMB NSAID Consort Diagram}
\usage{
dsmb_nsaid_consort_diagram(analytic, not_enrolled_other = NULL)
}
\arguments{
\item{analytic}{This is the analytic data set that must include screened, eligible, 
consented, randomized, enrolled, active, enrolled_discontinuation, refused, late_ineligible, and the meta construct column}

\item{not_enrolled_other}{is a meta construct that is NULL by default}

\item{completed_str}{is the text for the completed box that defaults to 'Completed 12-month visit'}
}
\value{
nothing
}
\description{
This function visualizes the categorical percentages of Study Status
for the NSAID study
}
\examples{
\dontrun{
dsmb_nsaid_consort_diagram()
}
}
