% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{enrollment_by_injury_and_site}
\alias{enrollment_by_injury_and_site}
\title{Enrollment of subjects for ankle and plateau injuries by each site}
\usage{
enrollment_by_injury_and_site(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, injury_type, enrolled, facilitycode}
}
\value{
nothing
}
\description{
This function visualizes the enrollment by each site for each injury_type
}
\examples{
\dontrun{
enrollment_by_injury_and_site()
}
}
