% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{enrollment_by_site_last_days_var_disc}
\alias{enrollment_by_site_last_days_var_disc}
\title{enrollment_by_site tobra and sextant (var discontinued)}
\usage{
enrollment_by_site_last_days_var_disc(
  analytic,
  days = 0,
  discontinued = "discontinued",
  discontinued_colname = "Discontinued",
  include_exclusive_safety_set = FALSE,
  average = FALSE,
  cumulative_data = TRUE
)
}
\arguments{
\item{analytic}{This is the analytic data set that must include screened, eligible, refused, not_consented, not_randomized, consented_and_randomized, enrolled, site_certified_days, 
facilitycode, screened_date, consented, randomized, consent_date, discontinued}

\item{days}{the number of last days to include in the last days summary section of the table}

\item{discontinued}{this is a meta construct where you can specify your discontinued construct like 'discontinued' or 'adjudicated_discontinued' (defaults to 'discontinued')}

\item{discontinued_colname}{this determines the label applied to the discontinued column of your choosing (defaults to 'Discontinued')}

\item{include_exclusive_safety_set}{this is a toggle that will include a exclusive_safety_set construct if you want it included (defaults to FALSE)}
}
\value{
An HTML table.
}
\description{
Visualizes the number of subjects enrolled, not enrolled etc, with parameter specifications to include more columns. 

For other enrollment by site visualizations that may better fit your study, please look at: enrollment_by_site, 
enrollment_by_site_last_days_var_disc, enrollment_status_by_site, enrollment_status_by_site_var_discontinued
}
\examples{
enrollment_by_site_last_days_var_disc("Replace with Analytic Tibble")

}
