% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{enrollment_by_site_last_days_var_disc_i}
\alias{enrollment_by_site_last_days_var_disc_i}
\title{enrollment_by_site_last_days_var_disc_i}
\usage{
enrollment_by_site_last_days_var_disc_i(
  analytic,
  days = 0,
  average = FALSE,
  cumulative_data = TRUE
)
}
\arguments{
\item{analytic}{Analytic data set.}

\item{days}{Number of last days to include.}

\item{average}{Return average over the time period.}

\item{cumulative_data}{Include final counts.}
}
\value{
An HTML table.
}
\description{
Visualizes the screening and eligibility status of subjects (Left half of original table).
Includes: Screened, Eligible, Refused, and Not Consented.
}
