% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{enrollment_by_site_last_days_var_disc_ii}
\alias{enrollment_by_site_last_days_var_disc_ii}
\title{enrollment_by_site_last_days_var_disc_ii}
\usage{
enrollment_by_site_last_days_var_disc_ii(
  analytic,
  discontinued = "discontinued",
  discontinued_colname = "Discontinued",
  include_exclusive_safety_set = FALSE,
  average = FALSE,
  cumulative_data = TRUE
)
}
\arguments{
\item{analytic}{Analytic data set.}

\item{discontinued}{Meta construct for discontinued.}

\item{discontinued_colname}{Label for the discontinued column.}

\item{include_exclusive_safety_set}{Toggle for exclusive_safety_set.}

\item{average}{Return average over the time period (Average Enrolled per week).}

\item{cumulative_data}{Include final counts.}
}
\value{
An HTML table.
}
\description{
Visualizes the right half of the enrollment table: Consented & Randomized, Discontinued, Enrolled, and Safety Set.
Consented & Randomized is formatted as a percentage of Eligible.
}
