% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{enrollment_by_site_var_disc}
\alias{enrollment_by_site_var_disc}
\title{enrollment_by_site (var discontinued)}
\usage{
enrollment_by_site_var_disc(analytic, days)
}
\arguments{
\item{analytic}{This is the analytic data set that must include screened, eligible, refused, not_consented, not_randomized, consented_and_randomized, enrolled, site_certified_days, 
facilitycode, adjudicated_discontinued, screened_date}
}
\value{
nothing
}
\description{
This function visualizes the number of subjects enrolled, not enrolled etc, with specs for last 14 days and average by week 
study.
}
\examples{
\dontrun{
enrollment_by_site_var_disc()
}
}
