% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{enrollment_status_by_site}
\alias{enrollment_status_by_site}
\title{Number of Subjects Screened, Eligible, Enrolled and Not Enrolled}
\usage{
enrollment_status_by_site(analytic)
}
\arguments{
\item{analytic}{analytic data set that must include screened, eligible, refused, consented, enrolled, 
not_consented, discontinued_pre_randomization, site_certification_data_metrc_date, facilitycode, late_ineligible}
}
\value{
html table
}
\description{
Visualizes the count of the study statuses for each site, as well as displaying the days the site 
has been certified.

For other enrollment by site visualizations that may better fit your study, please look at: enrollment_status_by_site_var_discontinued, 
enrollment_by_site_last_days_var_disc, enrollment_status_by_site
}
\examples{
enrollment_status_by_site("Replace with Analytic Tibble")

}
