% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{enrollment_status_by_site_consent_pre_screening}
\alias{enrollment_status_by_site_consent_pre_screening}
\title{Number of Subjects Screened, Eligible, and Enrolled, by Consented and with Pre-Screening (Variable Discontinued)}
\usage{
enrollment_status_by_site_consent_pre_screening(
  analytic,
  discontinued = "discontinued",
  discontinued_colname = "Discontinued",
  only_total = FALSE
)
}
\arguments{
\item{analytic}{This is the analytic data set that must include screened, 
eligible, refused, consented, enrolled, not_consented, site_certified_days, facilitycode,
consent_date, not_randomized}

\item{discontinued}{meta construct for discontinued}

\item{discontinued_colname}{column name for discontinued to appear in visualization like "Adjudicated Discontinued"}

\item{only_total}{hide all the site specific rows}
}
\value{
An HTML table.
}
\description{
Visualizes the totals of each include construct by site, split into among eligible and among consented,
with consented between the pre screening and the screening stage

For other enrollment by site visualizations that may better fit your study, refer to: enrollment_by_site, 
enrollment_by_site_last_days_var_disc, enrollment_status_by_site, enrollment_status_by_site_var_discontinued
}
\examples{
enrollment_status_by_site_consent_pre_screening("Replace with Analytic Tibble")

}
