% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{enrollment_status_by_site_consent_pre_screening_i}
\alias{enrollment_status_by_site_consent_pre_screening_i}
\title{enrollment_status_by_site_consent_pre_screening_i}
\usage{
enrollment_status_by_site_consent_pre_screening_i(analytic, only_total = FALSE)
}
\arguments{
\item{analytic}{Analytic data set. Must include: screened, eligible, ineligible, 
consented, facilitycode, site_certification_date, pre_screened, pre_eligible.}

\item{only_total}{hide all the site specific rows}
}
\value{
An HTML table.
}
\description{
Visualizes the totals of pre-screening, consent, intra-operative screening, and eligibility by site.
Structure: Pre-Screen -> Consent -> Intra-Op Screen -> Eligible.
}
