% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{enrollment_status_by_site_var_discontinued}
\alias{enrollment_status_by_site_var_discontinued}
\title{Number of Subjects Screened, Eligible, Enrolled and Not Enrolled (Variable Discontinued)}
\usage{
enrollment_status_by_site_var_discontinued(
  analytic,
  discontinued = "discontinued",
  discontinued_colname = "Discontinued",
  only_total = FALSE
)
}
\arguments{
\item{analytic}{This is the analytic data set that must include screened, 
eligible, refused, consented, enrolled, not_consented, site_certified_days, facilitycode,
consent_date, not_randomized}

\item{discontinued}{meta construct for discontinued}

\item{discontinued_colname}{column name for discontinued to appear in visualization like "Adjudicated Discontinued"}

\item{only_total}{hide all the site specific rows}
}
\value{
An HTML table.
}
\description{
This function visualizes the enrollment totals for each site
}
\examples{
enrollment_status_by_site_var_discontinued("Replace with Analytic Tibble")

}
