% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{enrollment_status_by_site_var_discontinued_i}
\alias{enrollment_status_by_site_var_discontinued_i}
\title{enrollment_status_by_site_var_discontinued_i}
\usage{
enrollment_status_by_site_var_discontinued_i(
  analytic,
  pre_screened = NULL,
  pre_screened_eligible = NULL,
  only_total = FALSE
)
}
\arguments{
\item{analytic}{Analytic data set. Must include: screened, eligible, refused, 
consented, not_consented, site_certification_date, facilitycode.}

\item{pre_screened}{Optional column name for pre-screened counts.}

\item{pre_screened_eligible}{Optional column name for pre-screened eligible counts.}

\item{only_total}{hide all the site specific rows}
}
\value{
An HTML table.
}
\description{
Visualizes the totals of screening, eligibility, and consent by site.
}
