% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{enrollment_status_by_site_var_discontinued_ii}
\alias{enrollment_status_by_site_var_discontinued_ii}
\title{enrollment_status_by_site_var_discontinued_ii}
\usage{
enrollment_status_by_site_var_discontinued_ii(
  analytic,
  discontinued = "discontinued",
  discontinued_colname = "Discontinued",
  only_total = FALSE
)
}
\arguments{
\item{analytic}{Analytic data set. Must include: eligible, consented, enrolled, 
randomized, facilitycode, discontinued (or custom name).}

\item{discontinued}{meta construct for discontinued}

\item{discontinued_colname}{column name for discontinued to appear in visualization like "Adjudicated Discontinued"}

\item{only_total}{hide all the site specific rows}
}
\value{
An HTML table.
}
\description{
Visualizes the totals of randomization, discontinuation, and enrollment by site.
Values are calculated as percentages of the "Consented" population.
}
