% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{expected_and_followup_visit}
\alias{expected_and_followup_visit}
\title{Expected visit status for 3 Months, 6 Months, and 12 Months followup}
\usage{
expected_and_followup_visit(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include followup_due_3mo, followup_due_6mo, 
followup_due_12mo, followup_status_3mo, followup_status_6mo, followup_status_12mo}
}
\value{
nothing
}
\description{
This function uses status constructs but treats early, late and complete as mutually exclusive.
Therefore, complete is renamed to "On Time" and all three of them combined to Complete.
}
\examples{
\dontrun{
expected_and_followup_visit()
}
}
