% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{expected_and_followup_visit_overall}
\alias{expected_and_followup_visit_overall}
\title{Expected visit status for Overall Followup}
\usage{
expected_and_followup_visit_overall(analytic, pretty_cols = c())
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, followup_data}
}
\value{
An HTML table.
}
\description{
Returns the counts of all the statuses of the Overall follow-up form at every follow-up periods. Notably,
this function does not separate the counts by site.
}
\examples{
expected_and_followup_visit_overall("Replace with Analytic Tibble")

}
