% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{expected_and_followup_visit_sextant}
\alias{expected_and_followup_visit_sextant}
\title{Expected visit status for 3 Months, 6 Months, and 12 Months followup for Sextant}
\usage{
expected_and_followup_visit_sextant(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, time_zero, followup_status_crf09_2wk, followup_status_crf09_3mo, followup_status_crf09_6mo, 
followup_status_crf09_12mo}
}
\value{
nothing

nothing
}
\description{
This function uses status constructs but treats early, late and complete as mutually exclusive.
Therefore, complete is renamed to "On Time" and all three of them combined to Complete.

This function onlt looks at the clinical followup form (CRF09) for tobra, designations may have to be looked at
}
\examples{
\dontrun{
expected_and_followup_visit_sextant()
}
Expected visit status for 3 Months, 6 Months, and 12 Months followup for tobra

\dontrun{
expected_and_followup_visit_tobra()
}
}
