% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{expected_followup_visits_nsaids}
\alias{expected_followup_visits_nsaids}
\title{Expected and followup visits for NSAIDs}
\usage{
expected_followup_visits_nsaids(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include threemonth_followup_complete, sixmonth_followup_complete, twelvemonth_followup_complete, threemonth_followup_incomplete,
sixmonth_followup_incomplete, twelvemonth_followup_incomplete, threemonth_followup_missing, 
sixmonth_followup_missing, twelvemonth_followup_missing, sixmonth_followup_late, threemonth_followup_late,
twelvemonth_followup_late, expected_3mo_followup, expected_6mo_followup, expected_12mo_followup}
}
\value{
nothing
}
\description{
This function visualizes the Expected and followup visits completeness, incompleteness, missingness, 
for NSAIDs
}
\examples{
\dontrun{
expected_followup_visits_nsaids()
}
}
