% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{followup_12mo_status_by_site_sextant}
\alias{followup_12mo_status_by_site_sextant}
\title{Followup 12 month status by site for Sextant}
\usage{
followup_12mo_status_by_site_sextant(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, eligible, enrolled, facilitycode, 
followup_expected_12mo, time_zero, followup_complete_crf12_12mo, followup_incomplete_crf12_12mo, 
followup_early_crf12_12mo, followup_late_crf12_12mo, followup_missing_crf12_12mo, 
followup_not_started_crf12_12mo, followup_status_crf14_crf15_12mo, followup_status_crf09_12mo}
}
\value{
nothing
}
\description{
This function visualizes 12 month followup status by site using CRF12(Clinical followup form),
PROMIS pain interference/intensity + CRF15(Survey version) or CRF14 + CRF15(Redcap version), and CRF08(Medical
Record Review) form
}
\examples{
\dontrun{
followup_12mo_status_by_site_sextant()
}
}
