% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{followup_12mo_status_by_site_tobra}
\alias{followup_12mo_status_by_site_tobra}
\title{Followup 12mo status by site for tobra}
\usage{
followup_12mo_status_by_site_tobra(analytic)
}
\arguments{
\item{analytic}{study_id, df_date, enrolled, facilitycode, followup_status_crf08_12mo, followup_status_crf09_12mo, followup_status_crf12_12mo, followup_status_bank_12mo, followup_status_comp_12mo}
}
\value{
nothing
}
\description{
This function visualizes 12mo followup status by site for Clinical followup form(crf09) and patient
medical record review(crf08) for tobra weekly report, and PROS from promis and workers comp
}
\examples{
\dontrun{
followup_12mo_status_by_site_tobra()
}
}
