% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{followup_2wk_status_by_site_sextant}
\alias{followup_2wk_status_by_site_sextant}
\title{Followup 2 week status by site for Sextant}
\usage{
followup_2wk_status_by_site_sextant(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, eligible, enrolled, facilitycode, 
followup_expected_2wk, followup_expected_3mo, dwc_date, followup_expected_6mo, followup_expected_12mo,
followup_complete_2wk_crf12, followup_complete_3mo_crf12, followup_complete_6mo_crf12, 
followup_complete_12mo_crf12, followup_incomplete_2wk_crf12, followup_incomplete_3mo_crf12, 
followup_incomplete_6mo_crf12, followup_incomplete_12mo_crf12, followup_early_2wk_crf12, 
followup_early_3mo_crf12, followup_early_6mo_crf12, followup_early_12mo_crf12, followup_late_2wk_crf12, 
followup_late_3mo_crf12, followup_late_6mo_crf12, followup_late_12mo_crf12, followup_missing_2wk_crf12, 
followup_missing_3mo_crf12, followup_missing_6mo_crf12, followup_missing_12mo_crf12,
followup_not_started_2wk_crf12, followup_not_started_3mo_crf12, followup_not_started_6mo_crf12, 
followup_not_started_12mo_crf12, followup_status_2wk_crf14_crf15}
}
\value{
nothing
}
\description{
This function visualizes followup status by site for Clinical followup form(crf12) and patient
reported outcome forms(CRF 14 & 15) for Sextant weekly report
}
\examples{
\dontrun{
followup_2wk_status_by_site_sextant()
}
}
