% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{followup_2wk_status_by_site_tobra}
\alias{followup_2wk_status_by_site_tobra}
\title{Followup 2 week status by site for tobra}
\usage{
followup_2wk_status_by_site_tobra(analytic)
}
\arguments{
\item{analytic}{study_id, df_date, enrolled, facilitycode, followup_status_crf08_2wk, followup_status_crf09_2wk}
}
\value{
nothing
}
\description{
This function visualizes 2 weeks followup status by site for Clinical followup form(crf09) and patient
medical record review(crf08) for tobra weekly report
}
\examples{
\dontrun{
followup_2wk_status_by_site_tobra()
}
}
