% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{followup_3mo_status_by_site_sextant}
\alias{followup_3mo_status_by_site_sextant}
\title{Followup 3 month status by site for Sextant}
\usage{
followup_3mo_status_by_site_sextant(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, eligible, enrolled, facilitycode, followup_expected_3mo, dwc_date, 
followup_complete_3mo_crf12, followup_incomplete_3mo_crf12, followup_early_3mo_crf12, followup_late_3mo_crf12, 
followup_missing_3mo_crf12, followup_not_started_3mo_crf12, followup_status_3mo_crf14_crf15}
}
\value{
nothing
}
\description{
This function visualizes 3 month followup status by site for Clinical followup form(crf12) and patient
reported outcome forms(CRF 14 & 15) for Sextant weekly report
}
\examples{
\dontrun{
followup_3mo_status_by_site_sextant()
}
}
