% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{followup_3mo_status_by_site_tobra}
\alias{followup_3mo_status_by_site_tobra}
\title{Followup 3mo status by site for tobra}
\usage{
followup_3mo_status_by_site_tobra(analytic)
}
\arguments{
\item{analytic}{study_id, df_date, enrolled, facilitycode, followup_status_crf08_3mo, followup_status_crf09_3mo}
}
\value{
nothing
}
\description{
This function visualizes 3mo followup status by site for Clinical followup form(crf09) and patient
medical record review(crf08) for tobra weekly report
}
\examples{
\dontrun{
followup_3mo_status_by_site_tobra()
}
}
