% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{followup_6mo_status_by_site_sextant}
\alias{followup_6mo_status_by_site_sextant}
\title{Followup 6 month status by site for Sextant}
\usage{
followup_6mo_status_by_site_sextant(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, eligible, enrolled, facilitycode, 
followup_expected_6mo, dwc_date, followup_complete_6mo_crf12, followup_incomplete_6mo_crf12, 
followup_early_6mo_crf12, followup_late_6mo_crf12, followup_missing_6mo_crf12, followup_not_started_6mo_crf12, 
followup_status_6mo_crf14_crf15}
}
\value{
nothing
}
\description{
This function visualizes 6 month followup status by site for Clinical followup form(crf12) and patient
reported outcome forms(CRF 14 & 15) for Sextant weekly report
}
\examples{
\dontrun{
followup_6mo_status_by_site_sextant()
}
}
