% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{followup_form_all_timepoints_by_site}
\alias{followup_form_all_timepoints_by_site}
\title{Followup Data Single Form All Timepoints By Site}
\usage{
followup_form_all_timepoints_by_site(analytic, form_selection = "Overall")
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, followup_data}
}
\value{
nothing
}
\description{
Returns the designated followup form status by site, for all timepoints
}
\examples{
\dontrun{
followup_form_all_timepoints_by_site()
}
}
