% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{followup_forms_all_timepoints}
\alias{followup_forms_all_timepoints}
\title{Followup Data Multiple Forms and All Timepoints}
\usage{
followup_forms_all_timepoints(
  analytic,
  forms = NULL,
  timepoints = NULL,
  vertical = TRUE
)
}
\arguments{
\item{analytic}{analytic data set that must include study_id, followup_data, facilitycode}

\item{forms}{followup forms to output, as found in the followup_data construct}

\item{timepoints}{timepoints to output, as found in the followup_data construct}

\item{vertical}{whether to arrange the output vertically by form or horizontally.}
}
\value{
An HTML table.
}
\description{
Returns all of the statuses of the given follow-up forms at all timepoints by site. Not specifying forms results
in all follow-up forms being in the visualization.

NOTE: THIS VISUALIZATION CAN BE VERY LARGE!
}
\examples{
followup_forms_all_timepoints("Replace with Analytic Tibble", forms = c('Form 3', 'Form 2'), timepoints = c('3 Month', '6 Month'))
followup_forms_all_timepoints("Replace with Analytic Tibble", vertical = FALSE)

}
