% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{followup_forms_at_timepoint_by_site}
\alias{followup_forms_at_timepoint_by_site}
\title{Followup Data Single Forms and Timepoint By Site}
\usage{
followup_forms_at_timepoint_by_site(analytic, timepoint, forms, names = NULL)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, followup_data}
}
\value{
nothing
}
\description{
Returns the designated followup form status by site
}
\examples{
\dontrun{
followup_forms_at_timepoint_by_site()
}
}
