% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{fracture_characteristics}
\alias{fracture_characteristics}
\title{Fracture Characteristics}
\usage{
fracture_characteristics(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, enrolled, fracture_type, injury_gustilo,}
}
\value{
An HTML table.
}
\description{
This function visualizes fracture characteristics, broken down by tibial plateau or pilon, 
and then closed or open fracture with tscherne grades and gustilo types respectively
}
\examples{
\dontrun{
fracture_characteristics()
}
}
