% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{hardware_duration_statistics_by_site}
\alias{hardware_duration_statistics_by_site}
\title{Hardware duration statistics by site}
\usage{
hardware_duration_statistics_by_site(analytic, delta = FALSE)
}
\arguments{
\item{analytic}{analytic data set that must include study_id, hardware_duration, hardware_delta, facilitycode}

\item{delta}{uses dates instead of datetimes}

\item{by_site}{breaks down hardware times by site}
}
\value{
html table
}
\description{
Returns a table of buckets for the duration hardware was applied to the patient, organized
by site.

Candidate for general visualization which interprets numeric constructs into buckets
}
\examples{
\dontrun{
hardware_duration_statistics_by_site("Replace with Analytic Tibble")
}
}
