% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{ineligibility_by_reasons}
\alias{ineligibility_by_reasons}
\title{Number of patients Ineligible by Top 5 reasons of Exclusion}
\usage{
ineligibility_by_reasons(
  analytic,
  pre_screened = FALSE,
  n_top_reasons = 5,
  only_total = FALSE
)
}
\arguments{
\item{analytic}{analytic data set that must include facilitycode, screened, ineligible, ineligibility_reasons}

\item{pre_screened}{when pre_screened is TRUE then uses pre-screening constructs}

\item{n_top_reasons}{is by default set to 5 but in case there are less than 5 reasons then as many 
columns would be reflected in the ineligibility table as reasons exist.}

\item{only_total}{hides site specific rows}
}
\value{
An HTML table.
}
\description{
Visualizes the counts of ineligibility reasons (NOT NUMBER OF INELIGIBLE PARTICIPANTS) by site. The 
function will display a user-specified number of reasons broken down, and then collate the rest into
an "Other Reasons" column. Included are two columns depicting numbers of screened and ineligible
study participants.
}
\examples{
ineligibility_by_reasons("Replace with Analytic Tibble")
ineligibility_by_reasons("Replace with Analytic Tibble", n_top_reasons = 3, only_total = TRUE)

}
