% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{ineligibility_reasons_info}
\alias{ineligibility_reasons_info}
\title{Ineligibility reasons info}
\usage{
ineligibility_reasons_info(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id,
facilitycode, screened, ineligible, and ineligibility_reasons}
}
\value{
visualization
}
\description{
\
The ineligibility reasons info function returns the counts of ineligibility
reasons specified in the ineligibility_reasons construct. The output visualization
is split up by site, with a total row. Note that the counts are for specific
reasons, not people, and so will not total the number of participants considered
ineligible.
}
\examples{
\dontrun{
top_5_ineligibility_reasons()
}
}
