% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{injury_ankle_plateau_characteristics}
\alias{injury_ankle_plateau_characteristics}
\title{Injury characteristics for OTA classification and Schatzker Type injuries}
\usage{
injury_ankle_plateau_characteristics(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include injury_type, ankle_ota_class, schatzker_type, enrolled}
}
\value{
nothing
}
\description{
This function visualizes the Injury characteristics for OTA classification and Schatzker Types for Ankle and Plateau
injuries
}
\examples{
\dontrun{
injury_ankle_plateau_characteristics()
}
}
