% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{injury_characteristics_by_alternate_constructs}
\alias{injury_characteristics_by_alternate_constructs}
\title{Injury Characteristics}
\usage{
injury_characteristics_by_alternate_constructs(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include enrolled, injury_classification_ankle_ao, injury_at_work, injury_in_battle, 
injury_in_blast, injury_date, injury_mechanism, injury_side, injury_classification_tscherne, injury_type}
}
\value{
nothing
}
\description{
This function visualizes the certain injury characteristics for study participants study injuries
}
\examples{
\dontrun{
injury_characteristics_by_alternate_constructs()
}
}
