% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{monitoring_required}
\alias{monitoring_required}
\title{Monitoring required}
\usage{
monitoring_required(analytic, large_sites = c(), min_pts = c(10, 25))
}
\arguments{
\item{analytic}{analytic data set that must include enrolled, facilitycode, consent_date}

\item{large_sites}{sites which use the upper designation of minimum participants required}

\item{min_pts}{a numeric vector of length 2 which contains the lower and upper throshold for monitoring}
}
\value{
An HTML table.
}
\description{
Simple function that returns whether a site has reached the number of enrolled required for monitoring
to be required.
}
\examples{
monitoring_required("Replace with Analytic Tibble", large_sites = c('AAA', 'AAB'))

}
