% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{nonunion_surgery_outcome}
\alias{nonunion_surgery_outcome}
\title{Nonunion surgery outcome}
\usage{
nonunion_surgery_outcome(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include enrolled, 
followup_due_3mo, followup_due_12mo, nonunion_90day,  nonunion_1yr}
}
\value{
nothing
}
\description{
This function visualizes the Nonunion surgery outcome
}
\examples{
\dontrun{
nonunion_surgery_outcome()
}
}
