% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{not_complete_sae_deviation_by_type_auto_categories}
\alias{not_complete_sae_deviation_by_type_auto_categories}
\title{Number of Non-Completing Participants, SAEs, and Protocol Deviations by type with AUTO Protocol Deviation Categorization}
\usage{
not_complete_sae_deviation_by_type_auto_categories(
  analytic,
  category_defaults = c("Safety", "Informed Consent", "Eligibility",
    "Protocol Implementation", "Other")
)
}
\arguments{
\item{analytic}{analytic data set that must include enrolled, not_expected_reason, not_completed_reason, 
not_completed, protocol_deviation_full_data, sae_count, consented}
}
\value{
An HTML table.
}
\description{
Visualizes the number of non-completions, not expected, and SAEs for enrolled participants and Protocol 
Deviations by type for consented participants. Amongst enrolled, counts instances of presence of the
construct not_completed_reason for Not Completed count, not_expected_reason for Not expected count,
and sae_count not being 0 for SAE count. Protocol deviation counts are extracted from the protocol_deviation_full_data
long file, where "Other. . ." values are truncated to "Other."

Categories of protocol deviations are separated by indentation.
}
\examples{
not_complete_sae_deviation_by_type_auto_categories("Replace with Analytic Tibble")

}
