% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{not_enrolled_reason}
\alias{not_enrolled_reason}
\title{Not enrolled reason}
\usage{
not_enrolled_reason(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, facilitycode, study_id, not_enrolled_reason, pre_screened_notes}
}
\value{
An HTML table.
}
\description{
This function visualizes list of study_ids who were are not enrolled, the reasons, and the screening notes
reasons
}
\examples{
not_enrolled_reason("Replace with Analytic Tibble")

}
