% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{outcome_by_id}
\alias{outcome_by_id}
\title{Visualize patient outcomes by ID over time}
\usage{
outcome_by_id(
  analytic,
  event_name,
  random_sample = NULL,
  facilitycodes = NULL,
  days_since_tz = 365,
  sort_by_fu_days = FALSE
)
}
\arguments{
\item{analytic}{analytic data set that must include study_id, facilitycode, events_data, outcome_data 
and time_zero}

\item{event_name}{specific event to track (will mark first occurrence specially)}

\item{random_sample}{optional integer to limit to a random sample of IDs}

\item{facilitycodes}{optional character vector to limit to a certain facilities}

\item{sort_by_fu_days}{optional boolean that sorts the html by the span from time_zero of the most
available data we have}

\item{days_since_dz}{optional numeric keyowrd argument to filter only for rows whose time_zero occured 
at least that many days ago}
}
\value{
An HTML string containing an image tag with the base64-encoded timeline visualization in PNG format.
}
\description{
Creates a timeline visualization for each patient showing events relative to time zero for SINGLE event outcomes.
}
\examples{
outcome_by_id("Replace with Analytic Tibble", "test_outcome")
outcome_by_id("Replace with Analytic Tibble", "test_outcome", sort_by_followup_days = TRUE)
outcome_by_id("Replace with Analytic Tibble", "test_outcome", random_sample = 50, facilitycodes = c('AAA', 'AAB'))

}
