% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{overall_complications}
\alias{overall_complications}
\title{Overall complications}
\usage{
overall_complications(
  analytic,
  relatedness = TRUE,
  WB = NULL,
  breakout_other = FALSE
)
}
\arguments{
\item{analytic}{analytic data set that must include study_id, hardware_duration, hardware_delta, facilitycode}

\item{relatedness}{includes that column}

\item{WB}{if the study is Weight Bearing}

\item{breakout_other}{If TRUE, replaces "Other" with "Other: [other_info]". Defaults to FALSE.}
}
\value{
html table
}
\description{
Returns a table of the overall complications first ordered by complication alphabetically, then by relatedness (starting with most related), 
then by severity (starting with most severe), each row is a unique combination of those items
}
