% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{overall_complications_closed}
\alias{overall_complications_closed}
\title{Overall complications closed}
\usage{
overall_complications_closed(analytic, relatedness = TRUE, WB = NULL)
}
\arguments{
\item{analytic}{analytic data set that must include study_id, treatment_arm, complication_data}

\item{relatedness}{includes that column}

\item{WB}{if the study is Weight Bearing}
}
\value{
html table
}
\description{
Returns a table of the overall complications separated by Treatment Arm (Group A vs Group B).
Ordered by complication alphabetically (Other at bottom), then by relatedness, then by severity.
}
