% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_tracker.R
\name{progress_tracker}
\alias{progress_tracker}
\title{Progress Tracker}
\usage{
progress_tracker(
  analytic,
  style,
  showCheckPoints,
  trackers,
  numerator_construct = NA,
  denominator_construct = NA,
  construct_tracker_name = NA,
  construct_units = NA
)
}
\arguments{
\item{analytic}{analytic data set that must include study_id, and the associated data set called progress_tracker (dummy column with all values set to TRUE)}

\item{style}{integers 1-5 indicating the type of style preffered for the progress indicator}

\item{trackers}{a comma separated string of particular tracker names, use "ALL" for all trackers, full name doesnt need to be given}

\item{numerator_construct}{the name of the construct you want to use as the numerator for your construct progress bar, 
if denominator is NA this construct must be only TRUE or FALSE so it can be its own denominator}

\item{denominator_construct}{optional denominator construct to allow the ratioing of one construct to another}

\item{construct_tracker_name}{name of the tracker for the given construct params}

\item{construct_units}{the "units" that are being mesured within the given tracker}

\item{show_check_points}{boolean to render the names of the tasks next to the progress bar}
}
\value{
An HTML string that renders the progress bar with the associated settings
}
\description{
Generates a progress bar/visual to display a percentage and item based tracker of what things for a given study must be done.

Currently only rendering for the IVAC data set
}
\examples{
progress_tracker(1, TRUE, "ALL")
progress_tracker(5, FALSE, "FirstTracker, Sec")

}
