% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{promis_stats_by_time}
\alias{promis_stats_by_time}
\title{PROMIS stats by time}
\usage{
promis_stats_by_time(analytic)
}
\arguments{
\item{analytic}{enrolled, promis_pf 6wk - 12mo constructs, promis_pain_interference 6wk - 12mo constructs}
}
\value{
An HTML table.
}
\description{
Returns stat data on promis scores, including sample size, mean, and standard deviation
}
\examples{
wbs_main_paper_aos("Replace with Analytic Tibble")

}
