% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{recruitment_source_statistics}
\alias{recruitment_source_statistics}
\title{recruitment_source_statistics}
\usage{
recruitment_source_statistics(analytic)
}
\value{
An HTML table.
}
\description{
This function outputs a table with the baseline statistics organized by all possible combinations of recruitment source and clinic.
This table is produced for TBI DSMB.
}
\examples{
recruitment_source_statistics("Replace with Analytic Tibble")

}
