% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{refusal_reasons_by_site}
\alias{refusal_reasons_by_site}
\title{Refusal reasons by each site}
\usage{
refusal_reasons_by_site(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include facilitycode, screened, refused, refused_reason}
}
\value{
nothing
}
\description{
This function visualizes the reasons of refusal, total refused and total screened by each site.
}
\examples{
\dontrun{
refusal_reasons_by_site()
}
}
