% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{reg_categories_table}
\alias{reg_categories_table}
\title{Categorizes participant Ethnicity based on their gender}
\usage{
reg_categories_table(analytic)
}
\arguments{
\item{analytic}{analytic data set that must include, race, ethnicity, and sex, and enrolled}
}
\value{
An HTML table.
}
\description{
Creates a table to model/compare the ethnicity, races, and genders of all particpants in a study
}
