% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{sites_days_certified}
\alias{sites_days_certified}
\title{Sites Days Certified}
\usage{
sites_days_certified()
}
\value{
nothing
}
\description{
Returns a table with all the sites and the number of days they
were certified for the study, calculated from a beginning date
}
\examples{
\dontrun{
sites_days_certified()
}
}
