% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{treatment_characteristics_sextant}
\alias{treatment_characteristics_sextant}
\title{Treatment characteristics_sextant}
\usage{
treatment_characteristics_sextant(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include local_antibiotic_at_dwc,
systemic_antibiotic_post_dwc, no_other_antibiotic_at_dwc}
}
\value{
nothing
}
\description{
This function visualizes the treatment characteristics per protocol and assignmnet for Sextant.
}
\examples{
\dontrun{
treatment_characteristics_sextant()
}
}
