% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{treatment_crossover_and_nonadherance}
\alias{treatment_crossover_and_nonadherance}
\title{Treatment crossover and nonadherance}
\usage{
treatment_crossover_and_nonadherance(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include facilitycode, eligible, enrolled, dwc_date, 
dwc_tobra, dwc_vanco}
}
\value{
nothing
}
\description{
This function visualizes the treatment crossover or any nonadherance occured during the Sextant 
study.
}
\examples{
\dontrun{
treatment_crossover_and_nonadherance()
}
}
