% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{treatment_crossover_and_nonadherence}
\alias{treatment_crossover_and_nonadherence}
\title{Treatment crossover and nonadherence}
\usage{
treatment_crossover_and_nonadherence(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include facilitycode, eligible, enrolled, time_zero, 
local_antibiotic_at_dwc}
}
\value{
nothing
}
\description{
This function visualizes the treatment crossover or any nonadherence occured during the Sextant 
study.
}
\examples{
\dontrun{
treatment_crossover_and_nonadherence()
}
}
