% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{visit_status_for_followup_by_form}
\alias{visit_status_for_followup_by_form}
\title{Expected, completed, missing, out of window visits by each form}
\usage{
visit_status_for_followup_by_form(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include mrr_status_6wk, mrr_status_3mo, mrr_status_6mo, 
mrr_status_12mo, cfu_status_6wk, cfu_status_3mo, cfu_status_6mo, cfu_status_12mo, pfu_status_6wk, pfu_status_3mo, 
pfu_status_6mo, pfu_status_12mo, bpi_status_6wk, bpi_status_3mo, bpi_status_6mo,bpi_status_12mo, aos_status_6wk, 
aos_status_3mo, aos_status_6mo, aos_status_12mo, koos_status_6wk, koos_status_3mo, koos_status_6mo, koos_status_12mo, 
sixweek_visit_expected, threemonth_visit_expected, sixmonth_visit_expected, twelvemonth_visit_expected, injury_type}
}
\value{
nothing
}
\description{
This function visualizes the expected visits for each timepoint for MRR, CFU, PFU, BPI, AOS, KOOS forms
}
\examples{
\dontrun{
visit_status_for_followup_by_form()
}
}
