% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{visit_status_for_followup_nsaids}
\alias{visit_status_for_followup_nsaids}
\title{Expected, completed, missing, out of window visits using crf08b form for NSAIDs}
\usage{
visit_status_for_followup_nsaids(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include crf08b_status_3mo, crf08b_status_6mo, 
crf08b_status_12mo, followup_due_3mo, followup_due_6mo, followup_due_12mo, enrolled}
}
\value{
nothing
}
\description{
This function visualizes the expected visits for each timepoint for CRF08B form
}
\examples{
\dontrun{
visit_status_for_followup_nsaids()
}
}
