% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{wbs_main_paper_all_characteristics}
\alias{wbs_main_paper_all_characteristics}
\title{Weight Bearing ALL characteristics for Main paper}
\usage{
wbs_main_paper_all_characteristics(analytic)
}
\arguments{
\item{analytic}{This is the analytic dataset that must include enrolled, injury_type,
sex, age, ethnicity_race, education_level,
preinjury_productive_activity, preinjury_work_demand, preinjury_work_hours,
tobacco_use, bmi, preinjury_health, insurance,
injury_gustilo, injury_classification_ankle_ota, soft_tissue_closure,
injury_mechanism, injury_randomization_days, pre_randomization_mobilization}
}
\value{
An HTML table.
}
\description{
This function outputs a table with the specified injury characteristics and patient characteristics
for enrolled patients with "Ankle" injuries. This table is produced for Weight bearing main paper.
}
\examples{
wbs_main_paper_all_characteristics("Replace with Analytic Tibble")

}
