% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{wbs_main_paper_aos}
\alias{wbs_main_paper_aos}
\title{Weight Bearing AOS for Main Paper}
\usage{
wbs_main_paper_aos(analytic)
}
\arguments{
\item{analytic}{enrolled, 
aos_disability_score_injured_leg_12mo, aos_disability_score_injured_leg_3mo, aos_disability_score_injured_leg_6mo, aos_disability_score_injured_leg_6wk, 
aos_pain_score_injured_leg_12mo, aos_pain_score_injured_leg_3mo, aos_pain_score_injured_leg_6mo, aos_pain_score_injured_leg_6wk, 
aos_score_injured_leg_12mo, aos_score_injured_leg_3mo, aos_score_injured_leg_6mo, aos_score_injured_leg_6wk}
}
\value{
An HTML table.
}
\description{
This function outputs a table with the AOS scores for enrolled patients. This table is produced for Weight bearing main paper.
}
\examples{
wbs_main_paper_aos("Replace with Analytic Tibble")

}
