% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{wbs_main_paper_bpi}
\alias{wbs_main_paper_bpi}
\title{Weight Bearing BPI for Main Paper}
\usage{
wbs_main_paper_bpi(analytic)
}
\arguments{
\item{analytic}{enrolled, 
bpi_severity_score_6wk, bpi_severity_score_3mo, bpi_severity_score_6mo, bpi_severity_score_12mo, 
bpi_interference_score_6wk, bpi_interference_score_3mo, bpi_interference_score_6mo, bpi_interference_score_12mo}
}
\value{
An HTML table.
}
\description{
This function outputs a table with the Brief Pain Inventory scores for enrolled patients. This table is produced for Weight bearing main paper.
}
\examples{
wbs_main_paper_bpi("Replace with Analytic Tibble")

}
