% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{wbs_main_paper_injury_characteristics}
\alias{wbs_main_paper_injury_characteristics}
\title{Weight Bearing Injury characteristics for Main paper}
\usage{
wbs_main_paper_injury_characteristics(analytic)
}
\arguments{
\item{analytic}{injury_classification_weber, injury_classification_open_close, 
injury_classification_lauge_hansen, injury_gustilo, injury_type, injury_classification_ankle_ota, 
definitive_fixation_construct, definitive_fixation_type, soft_tissue_closure, enrolled}
}
\value{
nothing
}
\description{
This function outputs a table with various injury characteristics for enrolled patients with "Ankle"
injuries. This table is produced for Weight bearing main paper.
}
\examples{
\dontrun{
wbs_main_paper_injury_characteristics()
}
}
