% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gray_tables.R
\name{adherence_by_site}
\alias{adherence_by_site}
\title{Adherence by site}
\usage{
adherence_by_site(analytic)
}
\arguments{
\item{analytic}{analytic data set that must include constructs facilitycode, df_date, enrolled, 
treatment_arm}
}
\value{
html table
}
\description{
Visualizes the status of patient study completion, organized by site. The returned 
table has 4 columns: Site, Enrolled, DWC complete, and adherence. The DWC complete
column is determined by the existence of a value for the df_date construct. Notably,
this visualization determines adherence based on the existence of a value for 
the treatment_arm construct.

A similar visualization is adherence_sextant, which uses multiple constructs to determine 
all types of adherence.
}
\examples{
adherence_by_site("Replace with Analytic Tibble")
}
