% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gray_tables.R
\name{adherence_by_site}
\alias{adherence_by_site}
\title{adherence by site}
\usage{
adherence_by_site(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include facilitycode, df_date}
}
\value{
nothing
}
\description{
This function visualizes the treatment crossover or any nonadherence occured during the study.
}
\examples{
adherence_by_site("Replace with Analytic Tibble")

}
