% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{amputation_characteristics_table}
\alias{amputation_characteristics_table}
\title{Amputation, Residual Limb and Prostheses Characteristics Summary Table}
\usage{
amputation_characteristics_table(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include: enrolled, 
amputation_days_since, prosthesis_days_per_week, prosthesis_hours_per_day, 
socket_comfort_score_sit, socket_comfort_score_stand, socket_comfort_score_walk, 
amputation_cause, ambulatory_device_list, medication_frequency_list, and meds_skin_list.}
}
\value{
An HTML table styled with kableExtra.
}
\description{
Visualizes the characteristics of enrolled subjects including years since amputation, 
cause, prosthesis use, ambulatory devices, socket comfort scores, and treatments.
}
\examples{
\dontrun{
}
}
