% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{amputations_and_gustilo_injury_characteristics}
\alias{amputations_and_gustilo_injury_characteristics}
\title{Amputations and gustilo injury characteristics}
\usage{
amputations_and_gustilo_injury_characteristics(analytic)
}
\arguments{
\item{analytic}{analytic data set that must include constructs enrolled, injury_gustilo_type, injury_amputation_status}
}
\value{
html table
}
\description{
Visualizes the injury characteristics for the amputation status (by category), and the fracture type 
(by gustilo). In order for the indents to work properly, the injury_gustilo_type 
construct must contain 7 unique values and the injury_amputation_status construct must contain 3 
unique values.
}
\examples{
amputations_and_gustilo_injury_characteristics("Replace with Analytic Tibble")

}
