% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{amputations_and_gustilo_injury_characteristics}
\alias{amputations_and_gustilo_injury_characteristics}
\title{Amputations and Gustilo Injury Characteristics}
\usage{
amputations_and_gustilo_injury_characteristics(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include enrolled, 
injury_gustilo_type, injury_amputation_status}
}
\value{
An HTML table.
}
\description{
This function visualizes the injury characteristics for amputations and Gustilo Injury types for 
Sextant study
}
\examples{
amputations_and_gustilo_injury_characteristics("Replace with Analytic Tibble")

}
