% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{ankle_and_plateau_x_ray_and_measurement_status}
\alias{ankle_and_plateau_x_ray_and_measurement_status}
\title{Ankle and Plateau X-Ray and Measurement Status}
\usage{
ankle_and_plateau_x_ray_and_measurement_status(analytic)
}
\arguments{
\item{analytic}{analytic data set that must include followup_expected_6wk,followup_expected_3mo,followup_expected_6mo, 
followup_expected_12mo,injury_type,followup_data,radiographs_taken_6wk, 
radiographs_taken_3mo,radiographs_taken_6mo,plat_tib_fib_overlap_6mo, 
plat_sagittal_pl_alignment_6mo, plat_patella_centered_6mo, plat_medial_prox_tibia_deg_6mo, 
plat_medial_lateral_diff_6mo,plat_condylar_width_6mo,plat_art_step_off_medial_6mo, 
plat_art_step_off_lateral_6mo,plat_femur_tibia_deg_6mo,plat_tib_fib_overlap_3mo, 
plat_sagittal_pl_alignment_3mo,plat_patella_centered_3mo,plat_medial_prox_tibia_deg_3mo, 
plat_medial_lateral_diff_3mo,plat_condylar_width_3mo,plat_art_step_off_medial_3mo, 
plat_art_step_off_lateral_3mo,plat_femur_tibia_deg_3mo,plat_tib_fib_overlap_6wk, 
plat_sagittal_pl_alignment_6wk,plat_patella_centered_6wk,plat_medial_prox_tibia_deg_6wk, 
plat_medial_lateral_diff_6wk,plat_condylar_width_6wk,plat_art_step_off_medial_6wk, 
plat_art_step_off_lateral_6wk,plat_femur_tibia_deg_6wk,ankle_talar_tilt_degrees_6wk, 
ankle_sagital_disp_6wk,ankle_coronal_plane_disp_6wk,ankle_talar_tilt_degrees_3mo, 
ankle_sagital_disp_3mo,ankle_coronal_plane_disp_3mo,ankle_talar_tilt_degrees_6mo, 
ankle_sagital_disp_6mo,ankle_coronal_plane_disp_6mo}
}
\value{
An HTML table.
}
\description{
This function visualizes Ankle and Plateau X-Ray and Measurement Status.

NOTE: this function is very unlikely to work with your study, check the construct requirements in
the analytic parameter.
}
\examples{
\dontrun{
ankle_and_plateau_x_ray_and_measurement_status()
}
}
