% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{baseline_characteristics_percent}
\alias{baseline_characteristics_percent}
\title{Baseline characteristics percent}
\usage{
baseline_characteristics_percent(
  analytic,
  sex = "sex",
  race = "ethnicity_race",
  education = "education_level",
  military = "military_status",
  sex_levels = c("Female", "Male", "Missing"),
  race_levels = c("Non-Hispanic White", "Non-Hispanic Black", "Hispanic", "Other",
    "Missing"),
  education_levels = c("Less than High School", "GED or High School Diploma",
    "More than High School", "Refused / Don't know", "Missing"),
  military_levels = c("Active Military", "Active Reserves", "Not Active Duty", "Missing")
)
}
\arguments{
\item{analytic}{This is the analytic data set that must include enrolled, age, age_group}

\item{sex}{is a meta construct that is required that defaults to "sex"}

\item{race}{is a meta construct that is required that defaults to "ethnicity_race"}

\item{education}{is a meta construct that is required that defaults to "education_level"}

\item{military}{is a meta construct that is required that defaults to "military_status"}

\item{sex_levels}{sets default values and orders for sex meta construct}

\item{race_levels}{sets default values and orders for race meta construct}

\item{education_levels}{sets default values and orders for education meta construct}

\item{military_levels}{sets default values and orders for military meta construct}
}
\value{
html table
}
\description{
Visualizes the categorical distribution of baseline characteristics 
sex, age, race, education, military, enrolled. See below as this is a generic visualization and includes meta construct for each of the 
analysis outputs. You may also specify the levels that these outputs have in the function call. 
Outputs two columns: type (sex, age, race, education, military), and their respective counts and percentages.
}
\examples{
baseline_characteristics_percent("Replace with Analytic Tibble")

}
