% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{certification_date_data}
\alias{certification_date_data}
\title{Status of IRB approvals and certification by site}
\usage{
certification_date_data(analytic, exclude_local_irb = FALSE)
}
\arguments{
\item{analytic}{This is the analytic data set that must include site_certified_date}

\item{exclude_local_irb}{defaults to False}
}
\value{
html table
}
\description{
Visualizes the sites for a given study and their dates of 
local, DOD, and METRC certifications. This function outputs 5 columns, Facility, Local or sIRB approval date, DoD approval date, 
certified by MCC to start screening, Number of days certified. To run this visualization a study needs qa site_certified_date long file.
}
\examples{
certification_date_data("Replace with Analytic Tibble")
certification_date_data("Replace with Analytic Tibble", exclude_local_irb = TRUE)

}
