% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{certification_date_data}
\alias{certification_date_data}
\title{Status of IRB Approvals and Certification by Site}
\usage{
certification_date_data(analytic, exclude_local_irb = FALSE)
}
\arguments{
\item{analytic}{This is the analytic data set that must include site_certified_date}

\item{exclude_local_irb}{defaults to False}
}
\value{
nothing
}
\description{
This function returns a list of sites and their dates of 
local, DOD, and METRC certifications
}
\examples{
\dontrun{
ineligibility_by_reasons()
}
}
