% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gray_tables.R
\name{characteristics_treatment}
\alias{characteristics_treatment}
\title{characteristics_treatment}
\usage{
characteristics_treatment(analytic)
}
\arguments{
\item{analytic}{analytic data set that must study_id, enrolled, df_date, plat_df_surgical_incision, 
pil_df_surgical_incision, df_number_procedures, adherence_to_intervention}
}
\value{
nothing
}
\description{
This function visualizes the treatment characteristics per protocol and assignment 
for tobra.
}
\examples{
characteristics_treatment("Replace with Analytic Tibble")

}
