% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_adherence_sextant}
\alias{closed_adherence_sextant}
\title{Closed adherence sextant}
\usage{
closed_adherence_sextant(analytic, footnotes = NULL)
}
\arguments{
\item{analytic}{This is the analytic data set that must include adherence_to_intervention_dwc,
adherence_to_intervention_post_dwc, adherence_to_no_other_antibiotic_dwc, treatment_arm}
}
\value{
An HTML table.
}
\description{
This function visualizes the treatment characteristics per protocol and assignment for Sextant for
each treatment group. 

NOTE: This is a closed version of adherence_sextant, request if adherence_sextant is present in your study. See adherence_sextant for more information.
}
\examples{
closed_adherence_sextant("Replace with Analytic Tibble")

}
