% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_amputations_and_gustilo_injury_characteristics}
\alias{closed_amputations_and_gustilo_injury_characteristics}
\title{Amputations and Gustilo injury characteristics, closed version}
\usage{
closed_amputations_and_gustilo_injury_characteristics(analytic)
}
\arguments{
\item{analytic}{analytic data set that must include enrolled, treatment_arm, injury_gustilo_type, 
injury_amputation_status}
}
\value{
An HTML table.
}
\description{
This is the closed version of amputations_and_gustilo_injury_characteristics; see its documentation
for details.
}
\examples{
closed_amputations_and_gustilo_injury_characteristics("Replace with Analytic Tibble")

}
