% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_appendix_A_SAEs}
\alias{closed_appendix_A_SAEs}
\title{Appendix A: Listing of Serious Adverse Events(SAEs) for closed report}
\usage{
closed_appendix_A_SAEs(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, sae_data}
}
\value{
An HTML table.
}
\description{
This function visualizes appendix for all the serious adverse events noted for the patients
}
