% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_appendix_B_deaths}
\alias{closed_appendix_B_deaths}
\title{Appendix B: Listing of any Death for closed report}
\usage{
closed_appendix_B_deaths(analytic)
}
\arguments{
\item{analytic}{analytic data set that must include constructs sae_data, death_date}
}
\value{
HTML.
}
\description{
Returns text describing sae data of all cases where the outcome is death. Death is determined by there
being a value for the death_date construct or the value Death in the sae_outcome column of the sae_data
long file.
}
\examples{
closed_appendix_B_deaths("Replace with Analytic Tibble")

}
