% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_appendix_C_not_expected_not_completed}
\alias{closed_appendix_C_not_expected_not_completed}
\title{Appendix C: Listing of any Not Expected and Not completed cases for closed report}
\usage{
closed_appendix_C_not_expected_not_completed(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, not_expected_data, not_completed_data}
}
\value{
HTML.
}
\description{
This function visualizes any not completedness and not expectedness occurred during the study time period.
}
\examples{
closed_appendix_C_not_expected_not_completed("Replace with Analytic Tibble")

}
