% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_appendix_D_protocol_deviation}
\alias{closed_appendix_D_protocol_deviation}
\title{Appendix D: Listing of any protocol deviations for closed report}
\usage{
closed_appendix_D_protocol_deviation(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, protocol_deviation_data}
}
\value{
An HTML table.
}
\description{
This function visualizes any protocol deviations occurred during the study time period.
}
\examples{
closed_appendix_D_protocol_deviation("Replace with Analytic Tibble")

}
