% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_baseline_characteristics_percent_nm}
\alias{closed_baseline_characteristics_percent_nm}
\title{Closed Baseline Characteristics Percent (No Military Status)}
\usage{
closed_baseline_characteristics_percent_nm(
  analytic,
  sex = "sex",
  race = "ethnicity_race",
  education = "education_level",
  sex_levels = c("Female", "Male", "Missing"),
  race_levels = c("Non-Hispanic White", "Non-Hispanic Black", "Hispanic", "Other",
    "Missing"),
  education_levels = c("Less than High School", "GED or High School Diploma",
    "More than High School", "Refused / Don't know", "Missing")
)
}
\arguments{
\item{analytic}{This is the analytic data set that must include treatment_arm enrolled, age, age_group}

\item{sex}{is a meta construct that is required that defaults to "sex"}

\item{race}{is a meta construct that is required that defaults to "race_ethnicity"}

\item{education}{is a meta construct that is required that defaults to "education_level"}

\item{sex_levels}{sets default values and orders for sex meta construct}

\item{race_levels}{sets default values and orders for race meta construct}

\item{education_levels}{sets default values and orders for education meta construct}
}
\value{
nothing
}
\description{
This function visualizes the categorical percentages of baseline characteristics sex, age, race, and education
}
\examples{
\dontrun{
closed_baseline_characteristics_percent_nm()
}
}
