% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_characteristics_treatment}
\alias{closed_characteristics_treatment}
\title{Closed Characteristics Treatment}
\usage{
closed_characteristics_treatment(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, enrolled, df_date, plat_df_surgical_incision, pil_df_surgical_incision, df_number_procedures, df_randomized_treatment, treatment_arm}
}
\value{
A kable table
}
\description{
This function visualizes definitive fixation by total df complete out of total enrolled, stages breakdown,
incisions broken down by plateau fractures and pilon fractures, and whether or not the study treatment adhered to protocol
by treatment arm.
}
\examples{
\dontrun{
closed_charcateristics_treatment(analytic)
}
}
