% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_characteristics_treatment}
\alias{closed_characteristics_treatment}
\title{closed characteristics_treatment}
\usage{
closed_characteristics_treatment(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must study_id, enrolled, df_date, plat_df_surgical_incision, 
pil_df_surgical_incision, df_number_procedures, adherence_to_intervention, treatment_arm}
}
\value{
An HTML table.
}
\description{
This function visualizes the treatment characteristics per protocol and assignment for tobra.
}
\examples{
\dontrun{
closed_characteristics_treatment()
}
}
