% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_complications_by_severity_relatedness}
\alias{closed_complications_by_severity_relatedness}
\title{Closed Complications by severity and relatedness}
\usage{
closed_complications_by_severity_relatedness(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, complication_data, treatment_arm}
}
\value{
html_table
}
\description{
visualizes the complications by severity and relatedness for dsmb report

This is a closed version of complications_by_severity_relatedness, request if complications_by_severity_relatedness is present in your study. 
See complications_by_severity_relatedness for more information.
}
\examples{
closed_complications_by_severity_relatedness("Replace with Analytic Tibble")

}
