% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_complications_overall}
\alias{closed_complications_overall}
\title{Closed complications reported overall from time_zero}
\usage{
closed_complications_overall(analytic, min_days = NULL, cutoff_days = NULL)
}
\arguments{
\item{analytic}{analytic data set that must include study_id, complication_data, treatment_arm}

\item{min_days}{keyword argument to pass in the number of days to get cut off date for complications}

\item{cutoff_days}{keyword argument to pass in the number of days to get cut off date for complications}
}
\value{
An HTML table.
}
\description{
Visualizes complication data and organizes it by treatment arm. Complications can be filtered to be
within a certain window using the min_days and cutoff_days parameters. See attached example table 
to see the format of the counts of complications.
}
\examples{
closed_complications_overall("Replace with Analytic Tibble")
closed_complications_overall("Replace with Analytic Tibble", min_days = 365, cutoff_days = 730)

}
