% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_figures.R
\name{closed_consort_diagram_nsaid_publication}
\alias{closed_consort_diagram_nsaid_publication}
\title{NSAID Consort Diagram: publication}
\usage{
closed_consort_diagram_nsaid_publication(
  analytic,
  outcome_day = 365,
  arm_a_str = "Group A",
  arm_b_str = "Group B"
)
}
\arguments{
\item{analytic}{analytic data set that must include
study_id, screened, ineligible, ineligibility_reasons, refused, constraint_48hrs,
constraint_admin, constraint_noconsent, constraint_other, non_enrolled_other, randomized,
adjudicated_inappropriate_enrollment, adjudicated_late_ineligible, surgery_or_healed_type,
surgery_or_healed_days, crossover, treatment_arm}

\item{outcome_day}{day at which the primary outcome status is assessed, defaults to 365}

\item{arm_a_str}{label for the Group A treatment arm, defaults to "Group A"}

\item{arm_b_str}{label for the Group B treatment arm, defaults to "Group B"}
}
\value{
An HTML string containing an image tag with the base64-encoded consort diagram in PNG format.
}
\description{
The closed version of consort_diagram_nsaid_publication, breaking down the flow of
patients after randomization by treatment_arm.

This consort diagram was made for the NSAID study, and so is unlikely to work for yours.
}
\examples{
closed_consort_diagram_nsaid_publication("Replace with Analytic Tibble")

}
