% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_enrollment_status_by_site}
\alias{closed_enrollment_status_by_site}
\title{Closed Number of Subjects Screened, Eligible, Enrolled and Not Enrolled}
\usage{
closed_enrollment_status_by_site(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include treatment_arm screened, 
eligible, refused, consented, enrolled, not_consented, early_withdraw, days_site_certified, 
facilitycode, late_ineligible, inappropriate_enrollment, late_refusal}
}
\value{
nothing
}
\description{
This function visualizes the enrollment totals for each site
}
\examples{
\dontrun{
closed_enrollment_status_by_site()
}
}
