% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_enrollment_status_by_site_var_discontinued}
\alias{closed_enrollment_status_by_site_var_discontinued}
\title{Closed Number of Subjects Screened, Eligible, Enrolled and Not Enrolled (Variable Discontinued)}
\usage{
closed_enrollment_status_by_site_var_discontinued(
  analytic,
  discontinued = "discontinued",
  discontinued_colname = "Discontinued"
)
}
\arguments{
\item{analytic}{This is the analytic data set that must include screened, 
eligible, refused, consented, enrolled, not_consented, site_certified_days, facilitycode, treatment_arm}

\item{discontinued}{meta construct for discontinued}

\item{discontinued_colname}{column name for discontinued to appear in visualization like "Adjudicated Discontinued"}
}
\value{
A kable table
}
\description{
This function visualizes the enrollment totals for each site by treatment arm
}
\examples{
\dontrun{
closed_enrollment_status_by_site_var_discontinued(analytic, discontinued = "adjudicated_discontinued", 
                                                  discontinued_colname = "Adjudicated Discontinued")
}
}
