% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_expected_and_followup_visit_overall}
\alias{closed_expected_and_followup_visit_overall}
\title{Expected visit status for Overall Followup}
\usage{
closed_expected_and_followup_visit_overall(analytic, footnotes = NULL)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, followup_data, treatment_arm}
}
\value{
html table
}
\description{
Visualizes the designated overall form(s) for a 
given study, as designated in the respective followup_data long file 
and organizes them by detected levels of followup periods. 

The function separates the data for the two study groups by treatment arm.

This is a closed version of expected_and_followup_visit_overall, request if expected_and_followup_visit_overall is present in your study. 
See expected_and_followup_visit_overall for more information.
}
\examples{
closed_expected_and_followup_visit_overall("Replace with Analytic Tibble")

}
