% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_followup_form_at_timepoint_by_site}
\alias{closed_followup_form_at_timepoint_by_site}
\title{Closed Followup Data Single Form and Timepoint By Site}
\usage{
closed_followup_form_at_timepoint_by_site(
  analytic,
  timepoint,
  form_selection,
  name = NULL
)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, followup_data, treatment_arm}

\item{timepoint}{the point in time to be considered in the visualization}

\item{form_selection}{the form to be considered in the visualization}

\item{name}{optional argument for changing the name of the followup form, for aesthetic use}
}
\value{
html table
}
\description{
Visualizes the designated followup form status across all sites, 
for a single timepoint and separated by the treatment_arm variable

This is a closed version of followup_form_at_timepoint_by_site, request if followup_form_at_timepoint_by_site is present in your study. 
See followup_form_at_timepoint_by_site for more information.
}
\examples{
closed_followup_form_at_timepoint_by_site("Replace with Analytic Tibble")

}
