% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_followup_forms_at_timepoint_by_site}
\alias{closed_followup_forms_at_timepoint_by_site}
\title{Closed Followup Data Multiple Forms and Single Timepoint By Site}
\usage{
closed_followup_forms_at_timepoint_by_site(
  analytic,
  timepoint,
  forms,
  names = NULL
)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, followup_data}

\item{timepoint}{the point in time to be considered in the visualization}

\item{forms}{the form to be considered in the visualization}

\item{names}{This is the analytic data set that must include study_id, followup_data}
}
\value{
nothing
}
\description{
Returns the designated followup form status by site
}
\examples{
\dontrun{
closex_followup_forms_at_timepoint_by_site()
}
}
