% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_fracture_characteristics}
\alias{closed_fracture_characteristics}
\title{Closed Fracture Characteristics}
\usage{
closed_fracture_characteristics(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, enrolled, fracture_type, injury_gustilo,
treatment_arm}
}
\value{
A kable table
}
\description{
This function visualizes fracture characteristics, broken down by tibial plateau or pilon,
and then closed or open fracture with tscherne grades and gustilo types respectively by treatment arm
}
\examples{
\dontrun{
closed_fracture_characteristics(analytic)
}
}
