% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_generic_characteristics}
\alias{closed_generic_characteristics}
\title{Closed Generic Characteristics}
\usage{
closed_generic_characteristics(
  analytic,
  constructs = c(),
  names_vec = c(),
  filter_cols = c("enrolled"),
  titlecase = FALSE,
  splits = NULL,
  subcategory_constructs = c(),
  bottom_order_levels = c("Missing"),
  mean_sd = c()
)
}
\arguments{
\item{analytic}{This is the analytic data set}

\item{constructs}{The constructs to run statistics from}

\item{names_vec}{The names of the constructs in the final visualization. Pass NA to attach a construct to the previous group's header without creating a new top border.}

\item{filter_cols}{The columns to filter the the data by (for totals and missing counts)}

\item{titlecase}{Changes construct values to Title Case}

\item{splits}{Splits the constructs if they are lists like "test_one,test_two" into two rows then counts them}

\item{subcategory_constructs}{This allows a characteristic to have a construct as a sub category, 
must be empty or specify a subcategory construct (or NA) for each construct (length of constructs == length of subcategory_constructs)}

\item{bottom_order_levels}{A vector of category names (e.g., "Missing", "Refused") to force to the bottom of the table, maintaining their order. Defaults to "Missing".}

\item{mean_sd}{A vector of construct names. If a construct is included here, it will be displayed as "Mean [SD]" with its calculated values, instead of categorical counts.}
}
\value{
html table
}
\description{
Visualizes basic count statistics for a number of constructs. Missing values are 
given the value "Missing."

This is a closed version of generic_characteristics, request if generic_characteristics is present in your study. 
See generic_characteristics for more information.
}
\examples{
\dontrun{
closed_generic_characteristics(analytic, constructs=c("age_cat", "age"), names_vec=c("Age", NA), mean_sd=c("age"))
}
}
