% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_generic_characteristics}
\alias{closed_generic_characteristics}
\title{Closed Generic Characteristics}
\usage{
closed_generic_characteristics(
  analytic,
  constructs = c(),
  names_vec = c(),
  filter_cols = c("enrolled"),
  titlecase = FALSE
)
}
\arguments{
\item{analytic}{This is the analytic data set}

\item{constructs}{The constructs to run statistics from}

\item{names_vec}{The names of the constructs in the final visualization}

\item{filter_cols}{The columns to filter the the data by (for totals and missing counts)}

\item{titlecase}{Changes construct values to Title Case}
}
\value{
nothing
}
\description{
Runs basic count statistics for a number of constructs. Missing values are 
given the value "Missing."
}
\examples{
\dontrun{
closed_generic_characteristics()
}
}
