% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_generic_characteristics}
\alias{closed_generic_characteristics}
\title{Closed Generic Characteristics}
\usage{
closed_generic_characteristics(
  analytic,
  constructs = c(),
  names_vec = c(),
  filter_cols = c("enrolled"),
  titlecase = FALSE,
  splits = NULL,
  subcategory_constructs = c()
)
}
\arguments{
\item{analytic}{This is the analytic data set}

\item{constructs}{The constructs to run statistics from}

\item{names_vec}{The names of the constructs in the final visualization}

\item{filter_cols}{The columns to filter the the data by (for totals and missing counts)}

\item{titlecase}{Changes construct values to Title Case}

\item{splits}{Splits the constructs if they are lists like "test_one,test_two" into two rows then counts them}

\item{subcategory_constructs}{This allows a characteristic to have a construct as a sub category, 
must be empty or specify a subcategory construct (or NA) for each construct (length of constructs == length of subcategory_constructs)}
}
\value{
An HTML table.
}
\description{
Runs basic count statistics for a number of constructs. Missing values are 
given the value "Missing."
}
\examples{
\dontrun{
generic_characteristics("Replace with Analytic Tibble", constructs="stages", names_vec="Stages")
}
}
