% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_ih_and_dc_crossover_monitoring_by_site_cutoff_date}
\alias{closed_ih_and_dc_crossover_monitoring_by_site_cutoff_date}
\title{Crossover Monitoring by Site since 01/01/2024}
\usage{
closed_ih_and_dc_crossover_monitoring_by_site_cutoff_date(
  analytic,
  footnotes = NULL
)
}
\arguments{
\item{analytic}{This is the analytic data set that must include enrolled, df_surg_completed, 
ih_discharge_date, crossover_inpatient, crossover_discharge, ih_discharge_date_on_time_zero, facilitycode, treatment_arm}
}
\value{
nothing
}
\description{
This function visualizes the crossovers by site in hospital and at discharge 
after 01/01/2021
}
\examples{
\dontrun{
closed_ih_and_dc_crossover_monitoring_by_site_cutoff_date()
}
}
