% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_injury_ankle_plateau_characteristics}
\alias{closed_injury_ankle_plateau_characteristics}
\title{Closed Injury characteristics for OTA classification and Schatzker Type injuries}
\usage{
closed_injury_ankle_plateau_characteristics(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include treatment_arm injury_type, ankle_ota_class, schatzker_type, enrolled}
}
\value{
nothing
}
\description{
This function visualizes the Injury characteristics for OTA classification and Schatzker Types for Ankle and Plateau
injuries
}
\examples{
\dontrun{
closed_injury_ankle_plateau_characteristics()
}
}
