% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_injury_characteristics_by_alternate_constructs}
\alias{closed_injury_characteristics_by_alternate_constructs}
\title{Closed Injury Characteristics}
\usage{
closed_injury_characteristics_by_alternate_constructs(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include enrolled, treatment_arm, injury_classification_ankle_ao, injury_at_work, injury_in_battle,
injury_in_blast, injury_date, injury_mechanism, injury_side, injury_classification_tscherne, injury_type}
}
\value{
A kable table
}
\description{
This function visualizes the certain injury characteristics for study participants study injuries by treatment arm
}
\examples{
\dontrun{
closed_injury_characteristics_by_alternate_constructs(analytic)
}
}
