% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_tables.R
\name{closed_not_complete_sae_deviation_by_type}
\alias{closed_not_complete_sae_deviation_by_type}
\title{Closed Number of Discontinued Participants, SAEs, and Protocol Deviations by type}
\usage{
closed_not_complete_sae_deviation_by_type(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include enrolled, not_expected_reason, not_completed_reason,
protocol_deviation_screen_consent, protocol_deviation_procedural, protocol_deviation_administrative, sae_count}
}
\value{
An HTML table.
}
\description{
This function visualizes the number of discontinuations, SAEs and Protocol Deviations by type
}
\examples{
\dontrun{
closed_not_complete_sae_deviation_by_type()
}
}
