% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{complications_by_severity_relatedness}
\alias{complications_by_severity_relatedness}
\title{Complications by severity and relatedness}
\usage{
complications_by_severity_relatedness(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, complication_data}
}
\value{
An HTML table.
}
\description{
This function visualizes the complications by severity and relatedness for dsmb report
}
\examples{
complications_by_severity_relatedness("Replace with Analytic Tibble")

}
