% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{consort_diagram}
\alias{consort_diagram}
\title{Consort Diagram}
\usage{
consort_diagram(
  analytic,
  final_period = "12 Month",
  definitive_event = "Definitive Fixation Complete",
  not_expected_adjudicated = TRUE
)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, screened, ineligible, eligible,
refused, consented, randomized, enrolled, time_zero, adjudicated_discontinued, completed, 
safety_set, exclusive_safety_set, not_completed, not_expected, active, missed_final_followup}

\item{final_period}{Defaults to 12 Month}

\item{definitive_event}{Event either DF or DWC}

\item{not_expected_adjudicated}{whether to note that the Not Expected was adjudicated}
}
\value{
nothing
}
\description{
This function visualizes the categorical percentages of study status as well as followup completions
}
\examples{
\dontrun{
consort_diagram()
}
}
