% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{consort_diagram}
\alias{consort_diagram}
\title{Consort Diagram}
\usage{
consort_diagram(analytic, definitive_event = "Definitive Fixation Complete")
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, screened, ineligible, eligible,
refused, consented, randomized, enrolled, time_zero, adjudicated_early_discontinued, followup_data, 
safety_set}

\item{definitive_event}{Event either DF or DWC}
}
\value{
nothing
}
\description{
This function visualizes the categorical percentages of study status as well as followup completions
}
\examples{
\dontrun{
consort_diagram()
}
}
