% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{consort_diagram_no_definitive_event}
\alias{consort_diagram_no_definitive_event}
\title{Consort Diagram No Definitive Event}
\usage{
consort_diagram_no_definitive_event(
  analytic,
  final_period = "12 Month",
  not_expected_adjudicated = TRUE
)
}
\arguments{
\item{analytic}{analytic data set that must include study_id, screened, ineligible, eligible,
refused, consented, randomized, enrolled, adjudicated_discontinued, completed, 
safety_set, exclusive_safety_set, not_completed, not_expected, active, missed_final_followup, incomplete_final_followup}

\item{final_period}{labels the final follow-up period}

\item{not_expected_adjudicated}{whether to note that the Not Expected was adjudicated}
}
\value{
An HTML string containing an image tag with the base64-encoded consort diagram in PNG format.
}
\description{
Visualizes study status data for studies without a definitive event. See functions consort_diagram,
dsmb_consort_diagram (all types) for related functions. The constructs completed, not_completed, 
missed_final_followup, and incomplete_final_followup are used in the last stage of the study, which
can be labelled by the final_period function parameter.
}
\examples{
consort_diagram_no_definitive_event("Replace with Analytic Tibble")

}
