% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{cumulative_enrolled}
\alias{cumulative_enrolled}
\title{Cumulative enrollment}
\usage{
cumulative_enrolled(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, enrolled, consent_date}
}
\value{
nothing
}
\description{
This function visualizes the Cumulative number of patients enrolled
}
\examples{
\dontrun{
cumulative_enrolled()
}
}
