% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{cumulative_enrolled}
\alias{cumulative_enrolled}
\title{Cumulative enrollment}
\usage{
cumulative_enrolled(analytic, bar_mode = FALSE)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, enrolled, consent_date}

\item{bar_mode}{set to TRUE to remove the line}
}
\value{
An HTML string containing an image tag with the base64-encoded consort diagram in PNG format.
}
\description{
This function visualizes the Cumulative number of patients enrolled
}
\examples{
\dontrun{
cumulative_enrolled()
}
}
