% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{cumulative_enrolled_los}
\alias{cumulative_enrolled_los}
\title{Cumulative enrollment for Length of Stay}
\usage{
cumulative_enrolled_los(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, ih_los_days}
}
\value{
An HTML string containing an image tag with the base64-encoded consort diagram in PNG format.
}
\description{
Visualizes the distribution of the number of days recorded in the ih_los_days construct.
}
\examples{
cumulative_enrolled_los("Replace with Analytic Tibble")

}
