% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{cumulative_percentage_ankle_injuries}
\alias{cumulative_percentage_ankle_injuries}
\title{Cumulative percentage for ankle injuries}
\usage{
cumulative_percentage_ankle_injuries(analytic)
}
\arguments{
\item{analytic}{analytic data set that must include study_id, injury_type (with an ankle value), 
enrolled}
}
\value{
An HTML string containing an image tag with the base64-encoded consort diagram in PNG format.
}
\description{
Visualizes the number of study participants over time, using the consent_date construct. Only ankle
injuries are accounted for.
}
\examples{
cumulative_percentage_ankle_injuries("Replace with Analytic Tibble")

}
