% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{cumulative_percentage_ankle_injuries}
\alias{cumulative_percentage_ankle_injuries}
\title{Cumulative percentage for ankle injuries}
\usage{
cumulative_percentage_ankle_injuries(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, injury_type, enrolled}
}
\value{
nothing
}
\description{
This function visualizes the Cumulative percentage for number of patients with ankle injuries over the period
of Year-Months(YYYY-MM) out of 526
}
\examples{
\dontrun{
cumulative_percentage_ankle_injuries()
}
}
