% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{cumulative_percentage_plateau_injuries}
\alias{cumulative_percentage_plateau_injuries}
\title{Cumulative percentage for Tibial Plateau injuries}
\usage{
cumulative_percentage_plateau_injuries(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, injury_type, enrolled}
}
\value{
An HTML string containing an image tag with the base64-encoded consort diagram in PNG format.
}
\description{
This function visualizes the Cumulative percentage for number of patients with Plateau injuries 
over the period of Year-Months(YYYY-MM, consent_date) out of 100
}
\examples{
\dontrun{
cumulative_percentage_plateau_injuries()
}
}
