% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{discrete_enrolled}
\alias{discrete_enrolled}
\title{Monthly Discrete Enrollment}
\usage{
discrete_enrolled(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, enrolled, consent_date}
}
\value{
An HTML string containing an image tag with the base64-encoded consort diagram in PNG format.
}
\description{
This function visualizes the discrete number of patients enrolled by month
}
\examples{
discrete_enrolled("Replace with Analytic Tibble")

}
