% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{discrete_enrolled}
\alias{discrete_enrolled}
\title{Monthly Discrete Enrollment}
\usage{
discrete_enrolled(analytic)
}
\arguments{
\item{analytic}{analytic data set that must include enrolled, consent_date}
}
\value{
An HTML string containing an image tag with the base64-encoded consort diagram in PNG format.
}
\description{
Visualizes the discrete number of participants enrolled by month using the consent_date construct. 
Notably, once a participant has been disenrolled, they are removed from this visualization.
}
\examples{
discrete_enrolled("Replace with Analytic Tibble")

}
