% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{dsmb_consort_diagram}
\alias{dsmb_consort_diagram}
\title{DSMB Consort Diagram}
\usage{
dsmb_consort_diagram(
  analytic,
  not_enrolled_other = NULL,
  completed_str = "Completed 12-month visit",
  late_ineligible = "late_ineligible",
  late_ineligible_str = "Late Ineligible",
  not_expected_adjudicated = FALSE
)
}
\arguments{
\item{analytic}{The analytic data set that must include the following columns: screened, eligible, consented, refused, discontinued_pre_randomization,
randomized, late_ineligible, enrolled, completed, not_completed, not_expected, active, missed_final_followup, incomplete_final_followup}

\item{not_enrolled_other}{A column in the dataset for cases that are eligible but not enrolled for reasons other than refusal (optional).}

\item{completed_str}{A string specifying the label for the completion status box. Defaults to "Completed 12-month visit".}

\item{late_inelgible}{defaults to late_ineligble but can be any construct like "adjudicated_discontinued"}

\item{late_inelgible_str}{defaults to Late Ineligible but can be any construct like "Adjudicated Discontinued"}
}
\value{
An HTML string containing an image tag with the base64-encoded consort diagram in PNG format.
}
\description{
This function visualizes the categorical percentages of Study Status for any study, similar to the NSAID consort diagram, but with customization endpoints.
}
\examples{
\dontrun{
dsmb_consort_diagram()
}
}
