% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{dsmb_consort_diagram}
\alias{dsmb_consort_diagram}
\title{DSMB Consort Diagram}
\usage{
dsmb_consort_diagram(
  analytic,
  not_enrolled_other = NULL,
  final_period = "12 Month",
  late_ineligible = "late_ineligible",
  late_ineligible_str = "Late Ineligible",
  not_expected_adjudicated = FALSE
)
}
\arguments{
\item{analytic}{analytic data set that must include the following columns: screened, eligible, 
consented, refused, discontinued_pre_randomization, randomized, late_ineligible, enrolled, completed, 
not_completed, not_expected, active, missed_final_followup, incomplete_final_followup}

\item{not_enrolled_other}{a column in the dataset for cases that are eligible but not enrolled for 
reasons other than refusal (optional, if not given is the number of eligible minus consented minus
refused).}

\item{final_period}{string specifying the label for the completion status box. Defaults to "12-month".}

\item{late_inelgible}{construct used for the count in the late ineligible box}

\item{late_inelgible_str}{label used for the late ineligible box}
}
\value{
An HTML string containing an image tag with the base64-encoded consort diagram in PNG format.
}
\description{
This function visualizes the categorical percentages of Study Status for any study, similar to the NSAID consort diagram, but with customization endpoints.

For other consort diagrams that may better fit your study, refer to: consort_diagram, consort_diagram_no_definitive_event, 
dsmb_consort_diagram, dsmb_consort_diagram_pre_no_def, dsmb_consort_diagram_pre_no_def_shifted_consent, 
dsmb_consort_diagram_pre_shifted_consent, dsmb_nsaid_consort_diagram.
}
\examples{
dsmb_consort_diagram("Replace with Analytic Tibble")

}
