% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{dsmb_consort_diagram_pre_no_def}
\alias{dsmb_consort_diagram_pre_no_def}
\title{DSMB Consort Diagram With Pre Screened and No Definitive Event}
\usage{
dsmb_consort_diagram_pre_no_def(
  analytic,
  final_period = "12 Month",
  adjudicated = FALSE
)
}
\arguments{
\item{analytic}{This is the analytic data set that must include pre_screened, pre_eligible, pre_ineligible, screened, eligible, ineligible,
consented, not_consented, randomized, enrolled, refused, completed, not_completed, not_expected, active, missed_final_followup, incomplete_final_followup}

\item{final_period}{Defaults to 12 Month}

\item{adjudicated}{whether to use adjudicated discontinued and not expected}
}
\value{
nothing
}
\description{
This function visualizes the categorical percentages of Study Status
for the NSAID study
}
\examples{
\dontrun{
dsmb_consort_diagram_pre_no_def()
}
}
