% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{dsmb_consort_diagram_pre_shifted_consent}
\alias{dsmb_consort_diagram_pre_shifted_consent}
\title{DSMB Consort Diagram With Pre Screened and No Definitive Event and with the Consented Group moved up}
\usage{
dsmb_consort_diagram_pre_shifted_consent(
  analytic,
  final_period = "12 Month",
  adjudicated = FALSE,
  definitive_event = "Nerve Surgery",
  pre_screened_name = NULL,
  screened_name = NULL
)
}
\arguments{
\item{analytic}{analytic data set that must include pre_screened, pre_eligible, screened, eligible,
consented, not_consented, randomized, enrolled, refused, completed, not_completed, not_expected, active, missed_final_followup, incomplete_final_followup}

\item{final_period}{visual option to name the last followup period}

\item{adjudicated}{visual option to show that discontinuation was adjudicated}

\item{definitive_event}{label for the definitive event}

\item{pre_screened_name}{label for pre-screened}

\item{screened_name}{label for screened}
}
\value{
An HTML string containing an image tag with the base64-encoded consort diagram in PNG format.
}
\description{
Visualizes all counts of study statuses.

Very similar to dsmb_consort_diagram_pre_no_def except the consented group is before screened in the
chain of study statuses.
}
\examples{
dsmb_consort_diagram_pre_shifted_consent("Replace with Analytic Tibble")
dsmb_consort_diagram_pre_shifted_consent("Replace with Analytic Tibble", definitive_event = 'TEST')

}
