% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{dsmb_nsaid_consort_diagram}
\alias{dsmb_nsaid_consort_diagram}
\title{DSMB NSAID Consort Diagram}
\usage{
dsmb_nsaid_consort_diagram(
  analytic,
  final_period = "12 Month",
  not_expected_adjudicated = TRUE
)
}
\arguments{
\item{analytic}{This is the analytic data set that must include screened, eligible, 
consented, not_consented, randomized, enrolled, refused, df_surg_completed, completed, 
not_completed, not_expected, active, missed_final_followup, incomplete_final_followup,
adjudicated_discontinued}

\item{final_period}{Defaults to 12 Month}

\item{not_expected_adjudicated}{whether to note that the Not Expected was adjudicated}

\item{definitive_event}{Event either DF or DWC}
}
\value{
An HTML string containing an image tag with the base64-encoded consort diagram in PNG format.
}
\description{
This function visualizes the categorical percentages of Study Status
for the NSAID study

For other consort diagrams that may better fit your study, refer to: consort_diagram, consort_diagram_no_definitive_event, 
dsmb_consort_diagram, dsmb_consort_diagram_pre_no_def, dsmb_consort_diagram_pre_no_def_shifted_consent, 
dsmb_consort_diagram_pre_shifted_consent, dsmb_nsaid_consort_diagram.
}
\examples{
dsmb_nsaid_consort_diagram("Replace with Analytic Tibble")

}
