% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{dsmb_nsaid_consort_diagram}
\alias{dsmb_nsaid_consort_diagram}
\title{DSMB NSAID Consort Diagram}
\usage{
dsmb_nsaid_consort_diagram(
  analytic,
  final_period = "12 Month",
  not_expected_adjudicated = TRUE
)
}
\arguments{
\item{analytic}{This is the analytic data set that must include screened, eligible, 
consented, not_consented, randomized, enrolled, refused, df_surg_completed, completed, not_completed, not_expected, active}

\item{final_period}{Defaults to 12 Month}

\item{not_expected_adjudicated}{whether to note that the Not Expected was adjudicated}

\item{definitive_event}{Event either DF or DWC}
}
\value{
nothing
}
\description{
This function visualizes the categorical percentages of Study Status
for the NSAID study
}
\examples{
\dontrun{
dsmb_nsaid_consort_diagram()
}
}
