% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{enrollment_and_followup_activities_overview}
\alias{enrollment_and_followup_activities_overview}
\title{Overview of enrollment and follow-up activities}
\usage{
enrollment_and_followup_activities_overview(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, followup_data,
facilitycode, screened, enrolled, eligible, screened_date}
}
\value{
visualization
}
\description{
Returns the screened, overall, and follow-up data, separated by sites
}
\examples{
\dontrun{
enrollment_and_followup_activities_overview()
}
}
