% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{enrollment_by_site}
\alias{enrollment_by_site}
\title{Enrollment of subjects by each site}
\usage{
enrollment_by_site(analytic, number_order = FALSE)
}
\arguments{
\item{analytic}{analytic data set that must include study_id, enrolled, facilitycode, consent_date}

\item{number_order}{arranges output by number of enrolled}
}
\value{
An HTML string containing an image tag with the base64-encoded consort diagram in PNG format.
}
\description{
This function visualizes the count of enrollment for each site.
}
\examples{
enrollment_by_site("Replace with Analytic Tibble")
enrollment_by_site("Replace with Analytic Tibble", number_order = TRUE)

}
