% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{enrollment_status_by_site}
\alias{enrollment_status_by_site}
\title{Number of Subjects Screened, Eligible, Enrolled and Not Enrolled}
\usage{
enrollment_status_by_site(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include screened, 
eligible, refused, consented, enrolled, not_consented, discontinued_pre_randomization, site_certified_days, 
facilitycode, late_ineligible}
}
\value{
nothing
}
\description{
This function visualizes the enrollment totals for each site
}
\examples{
\dontrun{
enrollment_status_by_site()
}
}
