% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{enrollment_status_by_site_var_discontinued}
\alias{enrollment_status_by_site_var_discontinued}
\title{Number of Subjects Screened, Eligible, Enrolled and Not Enrolled (Variable Discontinued)}
\usage{
enrollment_status_by_site_var_discontinued(
  analytic,
  discontinued = "discontined",
  discontinued_colname = "Discontinued"
)
}
\arguments{
\item{analytic}{This is the analytic data set that must include screened, 
eligible, refused, consented, enrolled, not_consented, days_site_certified, facilitycode}

\item{discontinued}{meta construct for discontinued}

\item{discontinued_colname}{column name for discontinued to appear in visualization like "Adjudicated Discontinued"}
}
\value{
nothing
}
\description{
This function visualizes the enrollment totals for each site
}
\examples{
\dontrun{
enrollment_status_by_site_var_discontinued()
}
}
