% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{enrollment_status_by_site_var_discontinued}
\alias{enrollment_status_by_site_var_discontinued}
\title{Number of Subjects Screened, Eligible, Enrolled and Not Enrolled (Variable Discontinued)}
\usage{
enrollment_status_by_site_var_discontinued(
  analytic,
  discontinued = "discontinued",
  discontinued_colname = "Discontinued",
  pre_screened = NULL,
  pre_screened_eligible = NULL,
  only_total = FALSE
)
}
\arguments{
\item{analytic}{This is the analytic data set that must include screened, 
eligible, refused, consented, enrolled, not_consented, site_certification_date, facilitycode,
consent_date, not_randomized}

\item{discontinued}{meta construct for discontinued}

\item{discontinued_colname}{column name for discontinued to appear in visualization like "Adjudicated Discontinued"}

\item{only_total}{hide all the site specific rows}
}
\value{
An HTML table.
}
\description{
Visualizes the totals of each include construct by site, split into among eligible and among consented.

For other enrollment by site visualizations that may better fit your study, refer to: enrollment_by_site, 
enrollment_by_site_last_days_var_disc, enrollment_status_by_site, enrollment_status_by_site_var_discontinued
}
\examples{
enrollment_status_by_site_var_discontinued("Replace with Analytic Tibble")

}
