% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{expected_and_followup_visit_overall}
\alias{expected_and_followup_visit_overall}
\title{Expected visit status for Overall Followup}
\usage{
expected_and_followup_visit_overall(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, followup_data}
}
\value{
nothing
}
\description{
This function only looks at the designated overall form(s) for a 
given study, as designated in the respective followup_data long file 
and organizes them by its detected levels of followup periods.
}
\examples{
\dontrun{
expected_and_followup_visit_overall()
}
}
