% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{followup_12mo_status_by_site_sextant}
\alias{followup_12mo_status_by_site_sextant}
\title{Followup 12 month status by site for Sextant}
\usage{
followup_12mo_status_by_site_sextant(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, eligible, enrolled, facilitycode, 
followup_expected_12mo, dwc_date, followup_complete_12mo_crf12, followup_incomplete_12mo_crf12, 
followup_early_12mo_crf12, followup_late_12mo_crf12, followup_missing_12mo_crf12, 
followup_not_started_12mo_crf12, followup_status_12mo_crf14_crf15, followup_status_12mo_crf09}
}
\value{
nothing
}
\description{
This function visualizes 12 month followup status by site using CRF12(Clinical followup form),
PROMIS pain interference/intensity + CRF15(Survey version) or CRF14 + CRF15(Redcap version), and CRF08(Medical
Record Review) form
}
\examples{
\dontrun{
followup_12mo_status_by_site_sextant()
}
}
