% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{followup_2wk_status_by_site_sextant}
\alias{followup_2wk_status_by_site_sextant}
\title{Followup 2 week status by site for Sextant}
\usage{
followup_2wk_status_by_site_sextant(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, eligible, enrolled, time_zero, facilitycode, 
followup_expected_2wk, followup_complete_crf12_2wk, followup_incomplete_crf12_2wk, followup_early_crf12_2wk, 
followup_late_crf12_2wk, followup_missing_crf12_2wk, followup_not_started_crf12_2wk, 
followup_status_crf14_crf15_2wk}
}
\value{
nothing
}
\description{
This function visualizes 2 weeks followup status by site for Clinical followup form(crf12) and patient
reported outcome forms(CRF 14 & 15) for Sextant weekly report
}
\examples{
\dontrun{
followup_2wk_status_by_site_sextant()
}
}
