% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{followup_6mo_status_by_site_sextant}
\alias{followup_6mo_status_by_site_sextant}
\title{Followup 6 month status by site for Sextant}
\usage{
followup_6mo_status_by_site_sextant(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, eligible, enrolled, facilitycode, 
followup_expected_6mo, time_zero, followup_complete_crf12_6mo, followup_incomplete_crf12_6mo, 
followup_early_crf12_6mo, followup_late_crf12_6mo, followup_missing_crf12_6mo, followup_not_started_crf12_6mo, 
followup_status_crf14_crf15_6mo}
}
\value{
nothing
}
\description{
This function visualizes 6 month followup status by site for Clinical followup form(crf12) and patient
reported outcome forms(CRF 14 & 15) for Sextant weekly report
}
\examples{
\dontrun{
followup_6mo_status_by_site_sextant()
}
}
