% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{followup_completion_time_stats}
\alias{followup_completion_time_stats}
\title{Follow-up Forms Time to Complete}
\usage{
followup_completion_time_stats(
  analytic,
  timepoints = c("6mo", "12mo"),
  ortho_timepoints = NULL,
  form_selection = "Overall"
)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, followup_data, event_time_zero,
and enrolled}

\item{timepoints}{the point in time to be considered in the visualization}

\item{form_selection}{the form to be considered in the visualization}
}
\description{
Returns summary statistics on the number of days to complete various follow-up forms.
}
\examples{
\dontrun{
followup_completion_time_stats()
}
}
