% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{followup_form_all_timepoints_by_site}
\alias{followup_form_all_timepoints_by_site}
\title{Followup Data Single Form All Timepoints By Site}
\usage{
followup_form_all_timepoints_by_site(
  analytic,
  form_selection = "Overall",
  included_columns = c("Not Expected", "Expected", "Complete", "Early", "Late", "Missed",
    "Not Started", "Incomplete")
)
}
\arguments{
\item{analytic}{analytic data set that must include study_id, followup_data}

\item{form_selection}{form whose statuses are to be investigated}

\item{included_colmns}{statuses to include in the vis}
}
\value{
An HTML table.
}
\description{
Returns the counts of the specified form statuses of a given form for all follow-up periods where that
form is present, by site. Specifying "Overall" in the form_selection results in a slightly more streamlined
look, without a header above the table.
}
\examples{
followup_form_all_timepoints_by_site("Replace with Analytic Tibble", form_selection = "Form 3")
followup_form_all_timepoints_by_site("Replace with Analytic Tibble", form_selection = "Form 3", 
  included_columns = c("Expected", "Complete", "Incomplete"))

}
