% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{followup_form_all_timepoints_by_site}
\alias{followup_form_all_timepoints_by_site}
\title{Followup Data Single Form All Timepoints By Site}
\usage{
followup_form_all_timepoints_by_site(
  analytic,
  form_selection = "Overall",
  included_columns = c("Not Expected", "Expected", "Complete", "Early", "Late", "Missed",
    "Not Started", "Incomplete")
)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, followup_data}

\item{form_selection}{The form to base the table on}

\item{included_colmns}{Defaults to c("Expected", "Complete", "Early", "Late", 'Missing', 'Not Started', 'Incomplete')}
}
\value{
An HTML table.
}
\description{
Returns the designated followup form status by site, for all timepoints
}
\examples{
followup_form_all_timepoints_by_site("Replace with Analytic Tibble", "2 Unit Period", "Form 3")

}
