% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{followup_form_at_timepoint_by_site}
\alias{followup_form_at_timepoint_by_site}
\title{Followup Data Single Form and Timepoint By Site}
\usage{
followup_form_at_timepoint_by_site(
  analytic,
  timepoint,
  form_selection,
  name = NULL
)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, followup_data, treatment_arm}

\item{timepoint}{the point in time to be considered in the visualization}

\item{form_selection}{the form to be considered in the visualization}

\item{name}{optional argument for changing the name of the followup form, for aesthetic use}
}
\value{
An HTML table.
}
\description{
Returns the designated followup form status across all sites, 
for a single timepoint and separated by the treatment_arm variable
}
\examples{
\dontrun{
followup_form_at_timepoint_by_site()
}
}
