% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{followup_forms_all_timepoints}
\alias{followup_forms_all_timepoints}
\title{Followup Data Multiple Forms and All Timepoints}
\usage{
followup_forms_all_timepoints(analytic, forms, timepoints)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, followup_data}

\item{forms}{followup forms to output, as found in the followup_data construct}

\item{timepoints}{timepoints to output, as found in the followup_data construct}
}
\value{
nothing
}
\description{
Returns the designated followup forms status at a specified range
of timepoints, not specified by site
}
\examples{
\dontrun{
followup_forms_all_timepoints()
}
}
