% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{followup_forms_at_timepoint_by_site}
\alias{followup_forms_at_timepoint_by_site}
\title{Followup Data Multiple Forms and Single Timepoints By Site}
\usage{
followup_forms_at_timepoint_by_site(analytic, timepoint, forms, names = NULL)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, followup_data}
}
\value{
An HTML table.
}
\description{
Visualizes the specified follow-up forms at a specific timeppoint, listed by site.

#' For other manipulations of this file that may better fit your study, please see: followup_completion_time_stats, 
followup_form_all_timepoints_by_site, followup_form_at_timepoint_by_site, followup_forms_all_timepoints, followup_forms_at_timepoint_by_site
}
\examples{
followup_forms_at_timepoint_by_site("Replace with Analytic Tibble", '3 Month', c('Form 3', 'Form 2'))

}
