% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{followup_forms_at_timepoint_by_site}
\alias{followup_forms_at_timepoint_by_site}
\title{Followup Data Multiple Forms and Single Timepoints By Site}
\usage{
followup_forms_at_timepoint_by_site(analytic, timepoint, forms, names = NULL)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, followup_data}
}
\value{
An HTML table.
}
\description{
Returns the designated followup forms status by site at a specifict
period
}
\examples{
followup_forms_at_timepoint_by_site("Replace with Analytic Tibble")

}
