% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{ih_and_dc_crossover_monitoring_by_site}
\alias{ih_and_dc_crossover_monitoring_by_site}
\title{Crossover Monitoring by Site}
\usage{
ih_and_dc_crossover_monitoring_by_site(analytic)

ih_and_dc_crossover_monitoring_by_site(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include enrolled, dfsurg_completed, 
ih_dischrg_date, ih_crossover, dc_crossover, ih_dischrg_date_on_time_zero, and facilitycode}
}
\value{
nothing

nothing
}
\description{
This function visualizes the crossovers by site in hospital and at discharge

This function visualizes the crossovers by site in hospital and at discharge
}
\examples{
\dontrun{
ih_and_dc_crossover_monitoring_by_site()
}
\dontrun{
ih_and_dc_crossover_monitoring_by_site()
}
}
