% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gray_tables.R
\name{ih_and_dc_crossover_monitoring_by_site}
\alias{ih_and_dc_crossover_monitoring_by_site}
\title{Crossover Monitoring by Site}
\usage{
ih_and_dc_crossover_monitoring_by_site(analytic)
}
\arguments{
\item{analytic}{analytic data set that must include enrolled, df_surg_completed, ih_discharge_date, 
crossover_inpatient, crossover_discharge, ih_discharge_date_on_time_zero, and facilitycode}
}
\value{
HTML table
}
\description{
Visualizes hospital discharges and patient crossovers out of the number of patients enrolled. Data
is organized per site.
}
\examples{
ih_and_dc_crossover_monitoring_by_site("Replace with Analytic Tibble")

}
