% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gray_tables.R
\name{ih_and_dc_crossover_monitoring_by_site_cutoff_date}
\alias{ih_and_dc_crossover_monitoring_by_site_cutoff_date}
\title{Crossover Monitoring by Site since 01/01/2024}
\usage{
ih_and_dc_crossover_monitoring_by_site_cutoff_date(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include enrolled, df_surg_completed, 
ih_discharge_date, crossover_inpatient, crossover_discharge, ih_discharge_date_on_time_zero, and facilitycode}
}
\value{
nothing
}
\description{
This function visualizes the crossovers by site in hospital and at discharge 
after 01/01/2021
}
\examples{
ih_and_dc_crossover_monitoring_by_site_cutoff_date("Replace with Analytic Tibble")

}
