% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{ineligibility_by_reasons}
\alias{ineligibility_by_reasons}
\title{Number of patients Ineligible by Top 5 reasons of Exclusion}
\usage{
ineligibility_by_reasons(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include facilitycode,  screened, ineligible, ineligibility_reasons}
}
\value{
nothing
}
\description{
This function visualizes the number of patients Ineligible by Top 5 reasons of Exclusion criteria
}
\examples{
\dontrun{
ineligibility_by_reasons()
}
}
