% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{monitoring_required}
\alias{monitoring_required}
\title{Monitoring required}
\usage{
monitoring_required(analytic, standard_threshold = 10, spec_threshold = list())
}
\arguments{
\item{analytic}{analytic data set that must include enrolled, facilitycode, consent_date}

\item{standard_threshold}{number to use as baseline threshold for monitoring requirement}

\item{spec_threshold}{optional kwarg that allows you to specify sites that have different monitoring
thresholds}
}
\value{
An HTML table.
}
\description{
Simple function that returns whether a site has reached the number of enrolled required for monitoring
to be required.
}
\examples{
monitoring_required("Replace with Analytic Tibble", spec_threshold = list('AAA' = 50), standard_threshold = 25)

}
