% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{nonunion_surgery_outcome}
\alias{nonunion_surgery_outcome}
\title{Nonunion surgery outcome}
\usage{
nonunion_surgery_outcome(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include enrolled, 
followup_expected_3mo, followup_expected_12mo, nonunion_90days,  nonunion_1yr}
}
\value{
An HTML table.
}
\description{
Visualizes the checks at 3 Months and 12 Months across the count of Non-Union at those timepoints.
}
\examples{
nonunion_surgery_outcome("Replace with Analytic Tibble")

}
