% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{not_enrolled_for_other_reasons}
\alias{not_enrolled_for_other_reasons}
\title{Not enrolled for other reasons}
\usage{
not_enrolled_for_other_reasons(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, facilitycode, able_to_participate, 
nonparticipation_text_given, constraint_noconsent, constraint_admin, constraint_other, constraint_other_txt, screened}
}
\value{
An HTML table.
}
\description{
This function visualizes list of study_ids who were screened howevere were not enrolled for "Other"
reasons
}
\examples{
\dontrun{
not_enrolled_for_other_reasons()
}
}
