% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{other_reason_refusal_by_site}
\alias{other_reason_refusal_by_site}
\title{Other reason of refusal by each site}
\usage{
other_reason_refusal_by_site(analytic)
}
\arguments{
\item{analytic}{analytic data set that must include study_id, facilitycode, screened_date, 
refused_reason_other}
}
\value{
html table
}
\description{
Returns a table of descriptions of all of the "Other" reasons given for refusal to participate.

See also the complementary table: refusal_reasons_by_site
}
\examples{
other_reason_refusal_by_site("Replace with Analytic Tibble")

}
