% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{other_reason_refusal_by_site}
\alias{other_reason_refusal_by_site}
\title{Other reason of refusal by each site}
\usage{
other_reason_refusal_by_site(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, facilitycode, screened_date, 
refused_reason_other}
}
\value{
An HTML table.
}
\description{
This function visualizes list of each "Other" reason of refusal, total screened by each site.
}
\examples{
\dontrun{
other_reason_refusal_by_site()
}
}
