% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{outcome_by_name_overall}
\alias{outcome_by_name_overall}
\title{Outcome by Name Overall}
\usage{
outcome_by_name_overall(analytic, days_since_tz = 365)
}
\arguments{
\item{analytic}{This is the analytic data set that must include study_id, outcome_data, and enrolled}

\item{days_since_dz}{optional numeric keyowrd argument to filter only for rows whose time_zero occured at leas that many days ago}
}
\value{
An HTML table.
}
\description{
Returns summary statistics on the number of days to complete various follow-up forms.
}
\examples{
outcome_by_name_overall("Replace with Analytic Tibble")

}
