% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{outcome_distribution}
\alias{outcome_distribution}
\title{Outcome distribution}
\usage{
outcome_distribution(
  analytic,
  scarq = FALSE,
  pretty_outcome_names = list(),
  heirarchy = c("death_outcome", "amputation", "reoperation_dssi",
    "operation_related_to_complication", "nonoperative_complication",
    "scarq_symptom_score_6mo", "scarq_appearance_score_6mo")
)
}
\arguments{
\item{analytic}{analytic data set that must include study_id, hardware_duration, hardware_delta, facilitycode}

\item{by_site}{breaks down hardware times by site}

\item{delta}{uses dates instead of datetimes}
}
\value{
html table
}
\description{
Returns a table of the percentage of each type of event being present.

For the iVAC study, an argument for the scarq scores is present, which will add the mean and standard
deviation of the appearance and symptom scores.
}
