% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{participants_w_complications}
\alias{participants_w_complications}
\title{Participants with complications}
\usage{
participants_w_complications(analytic)
}
\arguments{
\item{analytic}{analytic data set that must include study_id, complication_data}
}
\value{
html table
}
\description{
Returns statistics of the presence of one or more complication across the participants of the study.
Notably, this is numerically different from counting across complications themselves, as one participant
can have multiple complications
}
\examples{
participants_w_complications("Replace with Analytic Tibble")
}
