% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{pathogen_characteristics}
\alias{pathogen_characteristics}
\title{Pathogen Characteristics}
\usage{
pathogen_characteristics(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include dssi_data}
}
\value{
An HTML table styled with kableExtra.
}
\description{
Visualizes the breakdown of the dssi_data long file, uses helper constructs to get counts of everything.
}
\examples{
\dontrun{
}
}
