% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{patient_characteristics_table}
\alias{patient_characteristics_table}
\title{Patient Characteristics Summary Table}
\usage{
patient_characteristics_table(analytic)
}
\arguments{
\item{analytic}{This is the analytic data set that must include: enrolled, age, 
sex, ethnicity_race, education, preinjury_health, tobacco_use, and comorbidities_list.}
}
\value{
An HTML table styled with kableExtra.
}
\description{
Visualizes the patient characteristics for enrolled subjects, including Age, Sex, 
Race/Ethnicity, Education, Preinjury Health, Tobacco Use, and Comorbidities.
}
\examples{
\dontrun{
}
}
