% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{vislib_query_issues_per_site}
\alias{vislib_query_issues_per_site}
\title{Visualization Library: Issues per site}
\usage{
vislib_query_issues_per_site(analytic)
}
\value{
HTML graph.
}
\description{
Returns interactive HTML of the query status of all the sites. The data displays the current and two
week previous status of the queries to show a sense of progress. To turn on this function, one must
not only set up queries but also talk to the Analytic Team to move the queries into the Analytic 
Codebase.
}
\examples{
\dontrun{
vislib_query_issues_per_site()
}
}
