% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_figures.R
\name{vislib_query_issues_per_site_basic}
\alias{vislib_query_issues_per_site_basic}
\title{Visualization Library: Issues per site (Basic)}
\usage{
vislib_query_issues_per_site_basic(analytic)
}
\value{
html table
}
\description{
Visualizes the number of open and untouched issues per site,
determined by the status column in the query_database being set to "Detected".
NOTE: this is not a dsmb visualization

See also: vislib_query_issues_per_site
}
\examples{
\dontrun{
vislib_query_issues_per_site_basic()
}
}
