% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_tables.R
\name{wbs_main_paper_patient_characteristics}
\alias{wbs_main_paper_patient_characteristics}
\title{Weight Bearing Patient Characteristics for Main paper}
\usage{
wbs_main_paper_patient_characteristics(analytic)
}
\arguments{
\item{analytic}{enrolled, injury_type, sex, age, ethnicity_race, education_level, patient_reported_self_efficacy_6mo, 
patient_reported_self_efficacy_12mo, preinjury_productive_activity, preinjury_work_demand, 
preinjury_work_hours, tobacco_use, bmi, preinjury_health, insurance_type}
}
\value{
An HTML table.
}
\description{
This function outputs a table with various patient characteristics/demographics for enrolled 
patients with "Ankle" injuries. This table is produced for Weight bearing main paper.
}
\examples{
\dontrun{
wbs_main_paper_patient_characteristics()
}
}
